+++ 24c04680d54506a1f11eedc009691e3b.css
@@ -1,0 +1,1 @@
+body{margin:0}.card-header{background-color:#fff}.card-header a{color:#000;font-size:1.1em}.nav-item{text-align:center}iframe{margin-top:5px}h4{margin-left:25px;font-weight:700;margin-bottom:20px}h4,p{font-family:Catamaran,sans-serif}p{font-weight:500}a{font-family:Catamaran,sans-serif;font-weight:700}select{-webkit-appearance:none;-moz-appearance:none;appearance:none;background:url(b9a8e2330c0974ed9ae0c39a3ae46285.png) right 10px center/30px no-repeat #fff;padding-right:20px}body{font-size:14px;font-family:Helvetica Nueue,Arial,Verdana,sans-serif;background-color:#fff}iframe{width:100%}.bs-callout{margin:15px 0 10px;padding:20px 30px 5px 20px;border-left:5px solid #eee}.bs-callout-catalogue{border-color:#fccc30;background-color:#fafafa}.bs-callout .icon{float:left;font-size:1.5em;margin-top:-4px}.bs-callout h4,.bs-callout p{margin-left:35px}.dropdown-toggle{margin-top:10px;margin-bottom:10px;font-family:Catamaran,sans-serif;font-weight:500}.dropdown-toggle,.dropdown-toggle:focus,.dropdown-toggle:hover{background-color:#343a40;color:#fff}.cataloguepane{height:500px;overflow-y:scroll;margin-bottom:15px}.applist{margin-top:15px}.appinfo,.submittercatalogue{margin-top:-10px}.create{background-color:#a168c8}.connect{background-color:#4db6ac}.collaborate{background-color:#f06292}.curate{background-color:#fccc30}.capture{background-color:#4fc3f7}.captivate{background-color:#4b768d}.check{background-color:#dadfe1}.bs-callout-contribute{background-color:#fafafa;border-color:#a168c8}.bs-thanks{display:none;background-color:#fafafa;border-color:#4db6ac}.btn-block:hover{cursor:pointer}form input:focus,form textarea:focus{border-color:#a168c8!important;border-width:1px}.instruction{color:#90a4ad}form{margin-top:20px}.form-group label{font-family:Catamaran,sans-serif;font-weight:500}.submitbutton{margin-bottom:15px;background-color:#343a40;color:#fff}.submitbutton:hover{opacity:.9}#honeypot{display:none}.bs-callout-about{background-color:#fafafa;border-color:#4dd0e1;margin-bottom:15px}.bs-callout a{color:#4dd0e1}.creativecommons{margin-bottom:20px;margin-left:35px}.clearfix:after,.clearfix:before{content:" ";display:table}.clearfix:after{clear:both}.clearfix{*zoom:1}
