--- 3f8440a54235fc6e658427f94a1a3f02.css
@@ -1,1 +1,0 @@
-.titlebarfront{height:250px;padding-top:120px;display:-webkit-box;display:-ms-flexbox;display:flex;background-size:contain!important;background-repeat:no-repeat!important;background-position:50%!important}.smalllogo{width:40px;height:40px;margin-left:20px;-moz-border-radius:50px;-webkit-border-radius:50px;background-size:35px;background-position:50% 35%;background-repeat:no-repeat}.apptitle{color:#fff;margin-left:10px;padding-top:5px;font-family:Catamaran,sans-serif;width:60%}.cardflipper{height:50px;background-color:#00a5bd;border-top-left-radius:10px;border-top-right-radius:10px}.cardflipper,.nextcard{text-align:center;color:#fff;padding-top:15px;font-family:Catamaran,sans-serif;font-size:1.1em;cursor:pointer}.nextcard{height:55px;background-color:#353b40;border-bottom-left-radius:10px;border-bottom-right-radius:10px;border-top-style:solid;border-width:1px}.mainlogocontainer{max-width:650px;margin-left:auto;margin-right:auto}.details{height:160px;background-color:#343a40;color:#fff}hr{width:90%;background-color:#fff;margin-top:25px;margin-bottom:0}.mediabackground{background-image:url(43762566c02c0862fd12f54fedd7f391.png)}.videocontainer{max-width:560px!important;max-height:315px!important;margin-left:auto;margin-right:auto}h4{font-weight:700}.card-subtitle,h4{font-family:Catamaran,sans-serif}.card-subtitle{font-weight:500;margin-top:5px;color:#fff!important;margin-left:70px;padding-right:40px;line-height:1.2em}.card-text{line-height:1.7em;margin-left:10px;margin-right:10px}p{font-family:Catamaran,sans-serif;font-weight:500}.tool-link,a{font-family:Catamaran,sans-serif;font-weight:700}.tool-link{font-size:1.2em;margin-left:10px}.tags{font-size:20px}.blooms-pills-main{padding:30px 25px 0 5%;font-family:Catamaran,sans-serif}.blooms-pills-sub{padding:0 5%;font-family:Catamaran,sans-serif}.activity{text-transform:capitalize}.create{background-color:#ba68c8}.connect{background-color:#4dd0e1}.collaborate{background-color:#f06292}.curate{background-color:#ffb74d}.capture{background-color:#4fc3f7}.captivate{background-color:#4db6ac}.check,.sub{background-color:#90a4ae}.framework{display:-webkit-box;display:-ms-flexbox;display:flex;-ms-flex-wrap:wrap;flex-wrap:wrap;margin-bottom:-5px;margin-top:-5px}.thermometer{-ms-flex-preferred-size:30%;flex-basis:30%;text-align:center;padding-top:5px}.thermometer2,.thermometer3{margin-top:5px}.usestext{-ms-flex-preferred-size:70%;flex-basis:70%;padding-left:15px;padding-right:10px}.frameworkheading{display:-webkit-box;display:-ms-flexbox;display:flex;-webkit-box-orient:horizontal;-webkit-box-direction:normal;-ms-flex-direction:row;flex-direction:row;background-color:#353b40;color:#fff;padding-top:15px}.frameworktext1{text-align:center;-ms-flex-preferred-size:30%;flex-basis:30%;margin-left:15px}.frameworktext2{text-align:left;-ms-flex-preferred-size:70%;flex-basis:70%;margin-left:15px}.thermometer1,.usestext1{margin-top:5px}.usestext2,.usestext3{padding-top:5px}.submitter{display:-webkit-box;display:-ms-flexbox;display:flex}.submitterinfo{margin-top:20px;margin-left:10px;color:#fff}.submitterimg{text-align:center;margin-top:20px;margin-left:5%;font-size:1.75em}.submitterschool{color:#999}.ratings{display:block;text-align:center;background-color:#343a40;color:#fff;border-bottom-left-radius:10px;border-bottom-right-radius:10px}.ratings .btn{background-color:transparent;margin-bottom:5px}.card-footer{background-color:#00a5bd;color:#fff}.card-container{-webkit-perspective:800px;-o-perspective:800px;perspective:800px;margin:15px 0;height:800px}.noborder{border:none}.full-height{height:100%}.manual-flip .card{background:transparent}.manual-flip.hover .card{-webkit-transform:rotateY(180deg);transform:rotateY(180deg)}.app-card{-webkit-transition:-webkit-transform .5s;transition:-webkit-transform .5s;transition:transform .5s;transition:transform .5s,-webkit-transform .5s;-webkit-transform-style:preserve-3d;transform-style:preserve-3d;right:0}.app-card,.face{position:absolute;top:0;bottom:0;left:0}.face{-webkit-backface-visibility:hidden;backface-visibility:hidden;-webkit-box-shadow:0 0 5px 0 rgba(51,51,51,.5);box-shadow:0 0 5px 0 rgba(51,51,51,.5);border-radius:10px;width:100%}.front{z-index:2;-webkit-transform:rotateY(0deg);transform:rotateY(0deg);display:-webkit-box;display:-ms-flexbox;display:flex;-webkit-box-orient:vertical;-webkit-box-direction:normal;-ms-flex-direction:column;flex-direction:column}.back{-webkit-transform:rotateY(180deg);transform:rotateY(180deg)}.clearfix:after,.clearfix:before{content:" ";display:table}.clearfix:after{clear:both}.clearfix{*zoom:1}@media (max-width:300px){.titlebarfront{-webkit-box-orient:vertical;-webkit-box-direction:normal;-ms-flex-direction:column;flex-direction:column;height:280px;padding-top:75px;padding-left:10px}.apptitle{width:85%;margin-top:10px}.framework{-webkit-box-orient:vertical;-webkit-box-direction:normal;-ms-flex-direction:column;flex-direction:column;-webkit-box-flex:1;-ms-flex:1;flex:1}.details{height:250px}.submitterimg{display:none}}@media (max-width:190px){.details,.titlebarfront{height:320px}}
