+++ 45ce3f541db5c4cf2772da18010780cb.js
@@ -1,0 +1,67 @@
+require=function(r,e,n){function t(n,o){function i(r){return t(i.resolve(r))}function f(e){return r[n][1][e]||e}if(!e[n]){if(!r[n]){var c="function"==typeof require&&require;if(!o&&c)return c(n,!0);if(u)return u(n,!0);var l=new Error("Cannot find module '"+n+"'");throw l.code="MODULE_NOT_FOUND",l}i.resolve=f;var s=e[n]=new t.Module(n);r[n][0].call(s.exports,i,s,s.exports)}return e[n].exports}function o(r){this.id=r,this.bundle=t,this.exports={}}var u="function"==typeof require&&require;t.isParcelRequire=!0,t.Module=o,t.modules=r,t.cache=e,t.parent=u;for(var i=0;i<n.length;i++)t(n[i]);return t}({38:[function(require,module,exports) {
+"use strict";function t(){}function e(t){for(var e,n,r=1,i=arguments.length;r<i;r++)for(e in n=arguments[r])t[e]=n[e];return t}function n(t,e){e.appendChild(t)}function r(t,e,n){e.insertBefore(t,n)}function i(t){t.parentNode.removeChild(t)}function s(t,e){for(;t.nextSibling&&t.nextSibling!==e;)t.parentNode.removeChild(t.nextSibling)}function o(t){for(;t.previousSibling;)t.parentNode.removeChild(t.previousSibling)}function a(t){for(;t.nextSibling;)t.parentNode.removeChild(t.nextSibling)}function u(t,e,n){for(;t.nextSibling&&t.nextSibling!==e;)n.appendChild(t.parentNode.removeChild(t.nextSibling))}function l(t,e){for(;t.firstChild;)e.appendChild(t.firstChild)}function c(t,e){for(;t.nextSibling;)e.appendChild(t.nextSibling)}function h(t,e){for(var n=t.parentNode;n.firstChild!==t;)e.appendChild(n.firstChild)}function f(t){for(var e=0;e<t.length;e+=1)t[e]&&t[e].d()}function p(){return document.createDocumentFragment()}function d(t){return document.createElement(t)}function v(t){return document.createElementNS("http://www.w3.org/2000/svg",t)}function m(t){return document.createTextNode(t)}function g(){return document.createComment("")}function x(t,e,n){t.addEventListener(e,n,!1)}function _(t,e,n){t.removeEventListener(e,n,!1)}function b(t,e,n){t.setAttribute(e,n)}function y(t,e,n){t.setAttributeNS("http://www.w3.org/1999/xlink",e,n)}function w(t){for(var e=[],n=0;n<t.length;n+=1)t[n].checked&&e.push(t[n].__value);return e}function S(t){return""===t?void 0:+t}function k(t){for(var e=[],n=0;n<t.length;n+=1)e.push({start:t.start(n),end:t.end(n)});return e}function C(t){return Array.from(t.childNodes)}function N(t,e,n,r){for(var i=0;i<t.length;i+=1){var s=t[i];if(s.nodeName===e){for(var o=0;o<s.attributes.length;o+=1){var a=s.attributes[o];n[a.name]||s.removeAttribute(a.name)}return t.splice(i,1)[0]}}return r?v(e):d(e)}function R(t,e){for(var n=0;n<t.length;n+=1){var r=t[n];if(3===r.nodeType)return r.data=e,t.splice(n,1)[0]}return m(e)}function A(t,e){try{t.type=e}catch(t){}}function E(t,e,n){t.style.setProperty(e,n)}function T(t,e){for(var n=0;n<t.options.length;n+=1){var r=t.options[n];if(r.__value===e)return void(r.selected=!0)}}function O(t,e){for(var n=0;n<t.options.length;n+=1){var r=t.options[n];r.selected=~e.indexOf(r.__value)}}function j(t){var e=t.querySelector(":checked")||t.options[0];return e&&e.__value}function D(t){return[].map.call(t.querySelectorAll(":checked"),function(t){return t.__value})}function B(t){return t}function F(t,e,n,r,i,s){for(var o="{\n",a=0;a<=1;a+=16.666/r){o+=100*a+"%{"+s(t+n*i(a))+"}\n"}return o+"100% {"+s(e)+"}\n}"}function L(t){for(var e=5381,n=t.length;n--;)e=(e<<5)-e^t.charCodeAt(n);return e>>>0}function q(t,e,n,r,i,s){var o,a=n(e,r),u=a.duration||300,l=a.easing||B;if(a.css&&!I.stylesheet){var c=d("style");document.head.appendChild(c),I.stylesheet=c.sheet}return i&&(a.css&&a.delay&&(o=e.style.cssText,e.style.cssText+=a.css(0)),a.tick&&a.tick(0)),{t:i?0:1,running:!1,program:null,pending:null,run:function(t,e){var n={start:window.performance.now()+(a.delay||0),intro:t,callback:e};a.delay?this.pending=n:this.start(n),this.running||(this.running=!0,I.add(this))},start:function(n){t.fire(n.intro?"intro.start":"outro.start",{node:e}),n.a=this.t,n.b=n.intro?1:0,n.delta=n.b-n.a,n.duration=u*Math.abs(n.b-n.a),n.end=n.start+n.duration,a.css&&(a.delay&&(e.style.cssText=o),n.rule=F(n.a,n.b,n.delta,n.duration,l,a.css),I.addRule(n.rule,n.name="__svelte_"+L(n.rule)),e.style.animation=(e.style.animation||"").split(", ").filter(function(t){return t&&(n.delta<0||!/__svelte/.test(t))}).concat(n.name+" "+u+"ms linear 1 forwards").join(", ")),this.program=n,this.pending=null},update:function(t){var e=this.program;if(e){var n=t-e.start;this.t=e.a+e.delta*l(n/e.duration),a.tick&&a.tick(this.t)}},done:function(){var t=this.program;this.t=t.b,a.tick&&a.tick(this.t),a.css&&I.deleteRule(e,t.name),t.callback(),t=null,this.running=!!this.pending},abort:function(){a.tick&&a.tick(1),a.css&&I.deleteRule(e,this.program.name),this.program=this.pending=null,this.running=!1}}}Object.defineProperty(exports,"__esModule",{value:!0});var I={running:!1,transitions:[],bound:null,stylesheet:null,activeRules:{},add:function(t){this.transitions.push(t),this.running||(this.running=!0,requestAnimationFrame(this.bound||(this.bound=this.next.bind(this))))},addRule:function(t,e){this.activeRules[e]||(this.activeRules[e]=!0,this.stylesheet.insertRule("@keyframes "+e+" "+t,this.stylesheet.cssRules.length))},next:function(){this.running=!1;for(var t=window.performance.now(),e=this.transitions.length;e--;){var n=this.transitions[e];n.program&&t>=n.program.end&&n.done(),n.pending&&t>=n.pending.start&&n.start(n.pending),n.running?(n.update(t),this.running=!0):n.pending||this.transitions.splice(e,1)}if(this.running)requestAnimationFrame(this.bound);else if(this.stylesheet){for(e=this.stylesheet.cssRules.length;e--;)this.stylesheet.deleteRule(e);this.activeRules={}}},deleteRule:function(t,e){t.style.animation=t.style.animation.split(", ").filter(function(t){return t.slice(0,e.length)!==e}).join(", ")}};function M(){return Object.create(null)}function P(e){this.destroy=t,this.fire("destroy"),this.set=this.get=t,!1!==e&&this._fragment.u(),this._fragment.d(),this._fragment=this._state=null}function U(t){P.call(this,t),this.destroy=function(){console.warn("Component was already destroyed")}}function V(t,e){return t!=t?e==e:t!==e||t&&"object"==typeof t||"function"==typeof t}function G(t,e){return t!=t?e==e:t!==e}function X(t,e,n,r,i){for(var s in e)if(n[s]){var o=r[s],a=i[s],u=e[s];if(u)for(var l=0;l<u.length;l+=1){var c=u[l];c.__calling||(c.__calling=!0,c.call(t,o,a),c.__calling=!1)}}}function z(t,e){var n=t in this._handlers&&this._handlers[t].slice();if(n)for(var r=0;r<n.length;r+=1)n[r].call(this,e)}function H(t){return t?this._state[t]:this._state}function J(t,e){t._observers={pre:M(),post:M()},t._handlers=M(),t._bind=e._bind,t.options=e,t.root=e.root||t,t.store=t.root.store||e.store}function K(t,e,n){var r=n&&n.defer?this._observers.post:this._observers.pre;return(r[t]||(r[t]=[])).push(e),n&&!1===n.init||(e.__calling=!0,e.call(this,this._state[t]),e.__calling=!1),{cancel:function(){var n=r[t].indexOf(e);~n&&r[t].splice(n,1)}}}function Q(t,e,n){var r=(t=""+t).search(/[.[]/);if(r>-1){var i="The first argument to component.observe(...) must be the name of a top-level property";throw r>0&&(i+=", i.e. '"+t.slice(0,r)+"' rather than '"+t+"'"),new Error(i)}return K.call(this,t,e,n)}function W(t,e){if("teardown"===t)return this.on("destroy",e);var n=this._handlers[t]||(this._handlers[t]=[]);return n.push(e),{cancel:function(){var t=n.indexOf(e);~t&&n.splice(t,1)}}}function Y(t,e){return"teardown"===t?(console.warn("Use component.on('destroy', ...) instead of component.on('teardown', ...) which has been deprecated and will be unsupported in Svelte 2"),this.on("destroy",e)):W.call(this,t,e)}function Z(t){this._set(e({},t)),this.root._lock||(this.root._lock=!0,et(this.root._beforecreate),et(this.root._oncreate),et(this.root._aftercreate),this.root._lock=!1)}function $(t){var n=this._state,r={},i=!1;for(var s in t)this._differs(t[s],n[s])&&(r[s]=i=!0);i&&(this._state=e({},n,t),this._recompute(r,this._state),this._bind&&this._bind(r,this._state),this._fragment&&(X(this,this._observers.pre,r,this._state,n),this._fragment.p(r,this._state),X(this,this._observers.post,r,this._state,n)))}function tt(t){if("object"!=typeof t)throw new Error(this._debugName+".set was called without an object of data key-values to update.");this._checkReadOnly(t),Z.call(this,t)}function et(t){for(;t&&t.length;)t.shift()()}function nt(t,e){this._fragment.m(t,e)}function rt(){this._fragment&&this._fragment.u()}function it(t){return t&&"function"==typeof t.then}var st={},ot={},at={};function ut(){this.store._remove(this)}var lt={destroy:P,get:H,fire:z,observe:K,on:W,set:Z,teardown:P,_recompute:t,_set:$,_mount:nt,_unmount:rt,_differs:V},ct={destroy:U,get:H,fire:z,observe:Q,on:Y,set:tt,teardown:U,_recompute:t,_set:$,_mount:nt,_unmount:rt,_differs:V};exports.blankObject=M,exports.destroy=P,exports.destroyDev=U,exports._differs=V,exports._differsImmutable=G,exports.dispatchObservers=X,exports.fire=z,exports.get=H,exports.init=J,exports.observe=K,exports.observeDev=Q,exports.on=W,exports.onDev=Y,exports.set=Z,exports._set=$,exports.setDev=tt,exports.callAll=et,exports._mount=nt,exports._unmount=rt,exports.isPromise=it,exports.PENDING=st,exports.SUCCESS=ot,exports.FAILURE=at,exports.removeFromStore=ut,exports.proto=lt,exports.protoDev=ct,exports.appendNode=n,exports.insertNode=r,exports.detachNode=i,exports.detachBetween=s,exports.detachBefore=o,exports.detachAfter=a,exports.reinsertBetween=u,exports.reinsertChildren=l,exports.reinsertAfter=c,exports.reinsertBefore=h,exports.destroyEach=f,exports.createFragment=p,exports.createElement=d,exports.createSvgElement=v,exports.createText=m,exports.createComment=g,exports.addListener=x,exports.removeListener=_,exports.setAttribute=b,exports.setXlinkAttribute=y,exports.getBindingGroupValue=w,exports.toNumber=S,exports.timeRangesToArray=k,exports.children=C,exports.claimElement=N,exports.claimText=R,exports.setInputType=A,exports.setStyle=E,exports.selectOption=T,exports.selectOptions=O,exports.selectValue=j,exports.selectMultipleValue=D,exports.linear=B,exports.generateRule=F,exports.hash=L,exports.wrapTransition=q,exports.transitionManager=I,exports.noop=t,exports.assign=e;
+},{}],37:[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.Store=void 0;var e=require("./shared.js");function t(t,s){this._observers={pre:(0,e.blankObject)(),post:(0,e.blankObject)()},this._changeHandlers=[],this._dependents=[],this._computed=(0,e.blankObject)(),this._sortedComputedProperties=[],this._state=(0,e.assign)({},t),this._differs=s&&s.immutable?e._differsImmutable:e._differs}(0,e.assign)(t.prototype,{_add:function(e,t){this._dependents.push({component:e,props:t})},_init:function(e){for(var t={},s=0;s<e.length;s+=1){var r=e[s];t["$"+r]=this._state[r]}return t},_remove:function(e){for(var t=this._dependents.length;t--;)if(this._dependents[t].component===e)return void this._dependents.splice(t,1)},_sortComputedProperties:function(){var t,s=this._computed,r=this._sortedComputedProperties=[],i=(0,e.blankObject)();function n(e){if(t[e])throw new Error("Cyclical dependency detected");if(!i[e]){i[e]=!0;var o=s[e];o&&(t[e]=!0,o.deps.forEach(n),r.push(o))}}for(var o in this._computed)t=(0,e.blankObject)(),n(o)},compute:function(e,t,s){var r,i=this,n={deps:t,update:function(n,o,a){var h=t.map(function(e){return e in o&&(a=!0),n[e]});if(a){var p=s.apply(null,h);i._differs(p,r)&&(r=p,o[e]=!0,n[e]=r)}}};n.update(this._state,{},!0),this._computed[e]=n,this._sortComputedProperties()},get:e.get,observe:e.observe,onchange:function(e){this._changeHandlers.push(e);var t=this;return{cancel:function(){var s=t._changeHandlers.indexOf(e);~s&&t._changeHandlers.splice(s,1)}}},set:function(t){var s=this._state,r=this._changed={},i=!1;for(var n in t){if(this._computed[n])throw new Error("'"+n+"' is a read-only property");this._differs(t[n],s[n])&&(r[n]=i=!0)}if(i){this._state=(0,e.assign)({},s,t);for(var o=0;o<this._sortedComputedProperties.length;o+=1)this._sortedComputedProperties[o].update(this._state,r);for(o=0;o<this._changeHandlers.length;o+=1)this._changeHandlers[o](this._state,r);(0,e.dispatchObservers)(this,this._observers.pre,r,this._state,s);var a=this._dependents.slice();for(o=0;o<a.length;o+=1){var h=a[o],p={};i=!1;for(var d=0;d<h.props.length;d+=1){var c=h.props[d];c in r&&(p["$"+c]=this._state[c],i=!0)}i&&h.component.set(p)}(0,e.dispatchObservers)(this,this._observers.post,r,this._state,s)}}}),exports.Store=t;
+},{"./shared.js":38}],32:[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0});var t="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t};function e(){return{flipped:!1}}var a={rotateCard:function(){this.set({flipped:!this.get("flipped")})}};function r(){var t=this;this.store.observe("currentCard",function(e){setTimeout(function(){$(function(){$('[data-toggle="tooltip"]').tooltip()})},20),t.set({flipped:!1})})}function i(t,e){var a,r,i,c,_,b,y,N,x,k,w,L,T,M,C,H,E,S,O,R,j,q,A,F,U,$,B,P,z,D,G,I,J,K,Q,V,W,X,Y,Z,tt,et,at,rt,it,nt,st,ct,ot,dt,lt,ft,ut,mt,ht,pt,vt,gt,_t,bt,yt,Nt,xt,kt,wt,Lt,Tt,Mt,Ct,Ht,Et,St,Ot,Rt,jt,qt,At,Ft,Ut,$t,Bt,Pt,zt,Dt,Gt,It,Jt,Kt,Qt,Vt,Wt,Xt,Yt,Zt,te,ee,ae,re,ie,ne,se,ce=e.card.id,oe=e.card.colour,de=e.card.id,le=e.card.assets.cover.image,fe=e.card.assets.cover.styles,ue=e.card.id,me=e.card.assets.logo.image,he=e.card.assets.logo.styles,pe=e.card.name,ve=e.card.tagline,ge=e.card.description,_e=e.card.url,be=e.card.submitter.name,ye=e.card.submitter.title,Ne=e.card.framework.enhance,xe=e.card.framework.empower,ke=e.card.framework.extend;function we(e){t.rotateCard()}for(var Le=e.card.activities,Te=[],Me=0;Me<Le.length;Me+=1)Te[Me]=n(t,o({},e,{activity:Le[Me],activity_index:Me}));var Ce=e.card.uses,He=[];for(Me=0;Me<Ce.length;Me+=1)He[Me]=s(t,o({},e,{use:Ce[Me],use_index:Me}));function Ee(e){var a=t.get();t.store.loadRandomCard(a.card.id)}function Se(e){t.rotateCard()}function Oe(e){var a=t.get();t.store.loadRandomCard(a.card.id)}return{c:function(){a=d("div"),r=d("style"),i=l("."),c=l(ce),_=l(" .cardflipper, .nextcard:hover {\n            background-color: "),b=l(oe),y=l(";\n        }\n        ."),N=l(de),x=l(" .titlebarfront {\n            background: linear-gradient(\n                rgba(52, 58, 64, 0.9) 0%, \n                rgba(52, 58, 64, 0.9) 95%\n            ), url("),k=l(le),w=l(");\n            "),L=l(fe),T=l("\n        }\n        ."),M=l(ue),C=l(" .smalllogo {\n            background-image: url("),H=l(me),E=l(");\n            "),S=l(he),O=l("\n        }"),R=l("\n    "),j=d("div"),q=d("div"),A=d("div"),(F=d("div")).innerHTML='<div><i class="fas fa-redo-alt faa-spin animated-hover" aria-hidden="true"></i> Flip card</div>',U=l("\n                "),$=d("div"),B=d("a"),z=l("\n                    "),D=d("div"),G=d("h5"),I=l(pe),J=l("\n                    "),K=d("h6"),Q=l(ve),V=l("\n                "),W=d("div"),X=d("p"),Y=l(ge),Z=l("\n                    "),tt=d("a"),et=l(_e),rt=l("\n                "),it=d("div"),nt=d("div"),st=d("h5");for(var t=0;t<Te.length;t+=1)Te[t].c();ct=l("\n                            "),ot=d("i"),dt=l("\n                    "),lt=d("div");for(t=0;t<He.length;t+=1)He[t].c();ft=l(" \n                        "),ut=d("i"),mt=l("\n                    "),ht=d("div"),(pt=d("div")).innerHTML='<i class="far fa-user" aria-hidden="true"></i>',vt=l("\n                        "),gt=d("div"),_t=d("p"),(bt=d("b")).textContent="Submitted by",yt=d("br"),Nt=l(be),xt=d("span"),kt=l(" | "),wt=l(ye),Lt=l("\n                "),(Tt=d("div")).innerHTML='<div><i class="fas fa-magic" aria-hidden="true"></i>  Try another app</div>',Mt=l("\n\n            "),Ct=d("div"),(Ht=d("div")).innerHTML='<div><i class="fa faa-spin fa-repeat" aria-hidden="true"></i> Flip back</div>',Et=l("\n                "),St=d("div"),Ot=d("div"),Rt=d("div"),jt=d("iframe"),At=l("\n                "),(Ft=d("div")).innerHTML='<p class="frameworktext1">Level</p>\n                    <p class="frameworktext2">Example Activities</p>',Ut=l("\n                "),$t=d("div"),(Bt=d("div")).innerHTML='<h2><i class="fas fa-thermometer-quarter usesimage" aria-hidden="true"></i></h2>\n                        <p><strong>Enhance</strong></p>',Pt=l("\n                    "),zt=d("div"),Dt=d("p"),Gt=l(Ne),It=l("\n                    "),(Jt=d("div")).innerHTML='<h2><i class="fas fa-thermometer-half" aria-hidden="true"></i></h2>\n                        <p><strong>Empower</strong></p>',Kt=l("\n                    "),Qt=d("div"),Vt=d("p"),Wt=l(xe),Xt=l("\n                    "),(Yt=d("div")).innerHTML='<h2><i class="fas fa-thermometer-full usesimage" aria-hidden="true"></i></h2>\n                        <p><strong>Extend</strong></p>',Zt=l("\n                    "),te=d("div"),ee=d("p"),ae=l(ke),re=l("\n                "),(ie=d("div")).innerHTML='<div><i class="fas fa-magic" aria-hidden="true"></i>  Try another app</div>',this.h()},h:function(){r.type="text/css",F.className="cardflipper",f(F,"role","button"),u(F,"click",we),B.className="smalllogo faa-tada animated-hover",B.href=P=e.card.url,B.target="_blank",D.className="apptitle",$.className="titlebarfront",X.className="card-text",tt.className="tool-link",tt.href=at=e.card.url,tt.target="_blank",W.className="card-body",ot.className="fa fa-question-circle-o",ot.dataset.toggle="tooltip",ot.dataset.placement="top",ot.title="Activities this app can support",f(ot,"aria-hidden","true"),nt.className="pills-main",ut.className="fa fa-question-circle-o",ut.dataset.toggle="tooltip",ut.dataset.placement="top",ut.title="Examples of how this app can be used",f(ut,"aria-hidden","true"),lt.className="pills-sub",pt.className="submitterimg",xt.className="submitterschool",_t.className="small",gt.className="submitterinfo",ht.className="submitter",it.className="details",Tt.className="nextcard",f(Tt,"role","button"),u(Tt,"click",Ee),A.className="front face",Ht.className="cardflipper",f(Ht,"role","button"),u(Ht,"click",Se),jt.className="embed-responsive-item youtube",jt.src=qt=e.card.videoURL,f(jt,"frameborder","0"),jt.allowFullscreen=!0,Rt.className="embed-responsive embed-responsive-16by9",Ot.className="videocontainer",St.className="mediabackground",Ft.className="frameworkheading",Bt.className="thermometer thermometer1",zt.className="usestext usestext1",Jt.className="thermometer thermometer2",Qt.className="usestext usestext2",Yt.className="thermometer thermometer3",te.className="usestext usestext3",$t.className="card-body framework",ie.className="nextcard",f(ie,"role","button"),u(ie,"click",Oe),Ct.className="back face",q.className=ne="app-card card noborder "+e.card.id,j.className="card-container",a.className=se="manual-flip "+(e.flipped?"hover":"")},m:function(t,e){m(a,t,e),h(r,a),h(i,r),h(c,r),h(_,r),h(b,r),h(y,r),h(N,r),h(x,r),h(k,r),h(w,r),h(L,r),h(T,r),h(M,r),h(C,r),h(H,r),h(E,r),h(S,r),h(O,r),h(R,a),h(j,a),h(q,j),h(A,q),h(F,A),h(U,A),h($,A),h(B,$),h(z,$),h(D,$),h(G,D),h(I,G),h(J,D),h(K,D),h(Q,K),h(V,A),h(W,A),h(X,W),h(Y,X),h(Z,W),h(tt,W),h(et,tt),h(rt,A),h(it,A),h(nt,it),h(st,nt);for(var n=0;n<Te.length;n+=1)Te[n].m(st,null);h(ct,st),h(ot,st),h(dt,it),h(lt,it);for(n=0;n<He.length;n+=1)He[n].m(lt,null);h(ft,lt),h(ut,lt),h(mt,it),h(ht,it),h(pt,ht),h(vt,ht),h(gt,ht),h(_t,gt),h(bt,_t),h(yt,_t),h(Nt,_t),h(xt,_t),h(kt,xt),h(wt,xt),h(Lt,A),h(Tt,A),h(Mt,q),h(Ct,q),h(Ht,Ct),h(Et,Ct),h(St,Ct),h(Ot,St),h(Rt,Ot),h(jt,Rt),h(At,Ct),h(Ft,Ct),h(Ut,Ct),h($t,Ct),h(Bt,$t),h(Pt,$t),h(zt,$t),h(Dt,zt),h(Gt,Dt),h(It,$t),h(Jt,$t),h(Kt,$t),h(Qt,$t),h(Vt,Qt),h(Wt,Vt),h(Xt,$t),h(Yt,$t),h(Zt,$t),h(te,$t),h(ee,te),h(ae,ee),h(re,Ct),h(ie,Ct)},p:function(e,r){e.card&&ce!==(ce=r.card.id)&&(c.data=ce),e.card&&oe!==(oe=r.card.colour)&&(b.data=oe),e.card&&de!==(de=r.card.id)&&(N.data=de),e.card&&le!==(le=r.card.assets.cover.image)&&(k.data=le),e.card&&fe!==(fe=r.card.assets.cover.styles)&&(L.data=fe),e.card&&ue!==(ue=r.card.id)&&(M.data=ue),e.card&&me!==(me=r.card.assets.logo.image)&&(H.data=me),e.card&&he!==(he=r.card.assets.logo.styles)&&(S.data=he),e.card&&P!==(P=r.card.url)&&(B.href=P),e.card&&pe!==(pe=r.card.name)&&(I.data=pe),e.card&&ve!==(ve=r.card.tagline)&&(Q.data=ve),e.card&&ge!==(ge=r.card.description)&&(Y.data=ge),e.card&&_e!==(_e=r.card.url)&&(et.data=_e),e.card&&at!==(at=r.card.url)&&(tt.href=at);var i=r.card.activities;if(e.card){for(var d=0;d<i.length;d+=1){var l=o({},r,{activity:i[d],activity_index:d});Te[d]?Te[d].p(e,l):(Te[d]=n(t,l),Te[d].c(),Te[d].m(st,ct))}for(;d<Te.length;d+=1)Te[d].u(),Te[d].d();Te.length=i.length}var f=r.card.uses;if(e.card){for(d=0;d<f.length;d+=1){var u=o({},r,{use:f[d],use_index:d});He[d]?He[d].p(e,u):(He[d]=s(t,u),He[d].c(),He[d].m(lt,ft))}for(;d<He.length;d+=1)He[d].u(),He[d].d();He.length=f.length}e.card&&be!==(be=r.card.submitter.name)&&(Nt.data=be),e.card&&ye!==(ye=r.card.submitter.title)&&(wt.data=ye),e.card&&qt!==(qt=r.card.videoURL)&&(jt.src=qt),e.card&&Ne!==(Ne=r.card.framework.enhance)&&(Gt.data=Ne),e.card&&xe!==(xe=r.card.framework.empower)&&(Wt.data=xe),e.card&&ke!==(ke=r.card.framework.extend)&&(ae.data=ke),e.card&&ne!==(ne="app-card card noborder "+r.card.id)&&(q.className=ne),e.flipped&&se!==(se="manual-flip "+(r.flipped?"hover":""))&&(a.className=se)},u:function(){p(a);for(var t=0;t<Te.length;t+=1)Te[t].u();for(t=0;t<He.length;t+=1)He[t].u()},d:function(){v(F,"click",we),g(Te),g(He),v(Tt,"click",Ee),v(Ht,"click",Se),v(ie,"click",Oe)}}}function n(t,e){var a,r,i,n,s,c=e.activity,o=(e.activity_index,c.name);return{c:function(){a=d("span"),r=l(o),s=l(" "),this.h()},h:function(){a.className=i="badge activity "+c.name,a.dataset.toggle="tooltip",a.dataset.placement="top",a.title=n=c.description,f(a,"aria-hidden","true")},m:function(t,e){m(a,t,e),h(r,a),m(s,t,e)},p:function(t,e){c=e.activity,e.activity_index,t.card&&o!==(o=c.name)&&(r.data=o),t.card&&i!==(i="badge activity "+c.name)&&(a.className=i),t.card&&n!==(n=c.description)&&(a.title=n)},u:function(){p(a),p(s)},d:_}}function s(t,e){var a,r,i,n=e.use,s=(e.use_index,n);return{c:function(){a=d("span"),r=l(s),i=l(" "),this.h()},h:function(){a.className="badge sub"},m:function(t,e){m(a,t,e),h(r,a),m(i,t,e)},p:function(t,e){n=e.use,e.use_index,t.card&&s!==(s=n)&&(r.data=s)},u:function(){p(a),p(i)},d:_}}function c(t){b(this,t),this._state=o(e(),t.data);var a=r.bind(this);t.root||(this._oncreate=[]),this._fragment=i(this,this._state),this.root._oncreate.push(a),t.target&&(this._fragment.c(),this._fragment.m(t.target,t.anchor||null),y(this._oncreate))}function o(t){for(var e,a,r=1,i=arguments.length;r<i;r++)for(e in a=arguments[r])t[e]=a[e];return t}function d(t){return document.createElement(t)}function l(t){return document.createTextNode(t)}function f(t,e,a){t.setAttribute(e,a)}function u(t,e,a){t.addEventListener(e,a,!1)}function m(t,e,a){e.insertBefore(t,a)}function h(t,e){e.appendChild(t)}function p(t){t.parentNode.removeChild(t)}function v(t,e,a){t.removeEventListener(e,a,!1)}function g(t){for(var e=0;e<t.length;e+=1)t[e]&&t[e].d()}function _(){}function b(t,e){t._observers={pre:S(),post:S()},t._handlers=S(),t._bind=e._bind,t.options=e,t.root=e.root||t,t.store=t.root.store||e.store}function y(t){for(;t&&t.length;)t.shift()()}function N(t){this.destroy=_,this.fire("destroy"),this.set=this.get=_,!1!==t&&this._fragment.u(),this._fragment.d(),this._fragment=this._state=null}function x(t){return t?this._state[t]:this._state}function k(t,e){var a=t in this._handlers&&this._handlers[t].slice();if(a)for(var r=0;r<a.length;r+=1)a[r].call(this,e)}function w(t,e,a){var r=a&&a.defer?this._observers.post:this._observers.pre;return(r[t]||(r[t]=[])).push(e),a&&!1===a.init||(e.__calling=!0,e.call(this,this._state[t]),e.__calling=!1),{cancel:function(){var a=r[t].indexOf(e);~a&&r[t].splice(a,1)}}}function L(t,e){if("teardown"===t)return this.on("destroy",e);var a=this._handlers[t]||(this._handlers[t]=[]);return a.push(e),{cancel:function(){var t=a.indexOf(e);~t&&a.splice(t,1)}}}function T(t){this._set(o({},t)),this.root._lock||(this.root._lock=!0,y(this.root._beforecreate),y(this.root._oncreate),y(this.root._aftercreate),this.root._lock=!1)}function M(t){var e=this._state,a={},r=!1;for(var i in t)this._differs(t[i],e[i])&&(a[i]=r=!0);r&&(this._state=o({},e,t),this._recompute(a,this._state),this._bind&&this._bind(a,this._state),this._fragment&&(O(this,this._observers.pre,a,this._state,e),this._fragment.p(a,this._state),O(this,this._observers.post,a,this._state,e)))}function C(t,e){this._fragment.m(t,e)}function H(){this._fragment&&this._fragment.u()}function E(e,a){return e!=e?a==a:e!==a||e&&"object"===(void 0===e?"undefined":t(e))||"function"==typeof e}function S(){return Object.create(null)}function O(t,e,a,r,i){for(var n in e)if(a[n]){var s=r[n],c=i[n],o=e[n];if(o)for(var d=0;d<o.length;d+=1){var l=o[d];l.__calling||(l.__calling=!0,l.call(t,s,c),l.__calling=!1)}}}o(c.prototype,a,{destroy:N,get:x,fire:k,observe:w,on:L,set:T,teardown:N,_set:M,_mount:C,_unmount:H,_differs:E}),c.prototype._recompute=_,exports.default=c;
+},{}],33:[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0});var t="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t};function e(t,e){return"all"===e?t:t.filter(function(t){return-1!==t.activities.map(function(t){return t.name}).indexOf(e)})}function i(){return{filter:"all"}}function n(t,e){var i,n,r,s,o,m,g,y,b,x,C,N,k,w,S,O,j,E,L,P,T,$,B,M,A,F,H,q;function z(e){t.set({filter:e.target.value})}for(var D=e.filteredCards,G=[],I=0;I<D.length;I+=1)G[I]=a(t,l({},e,{card:D[I],card_index:I}));return{c:function(){i=c("div"),n=c("div"),(r=c("div")).innerHTML='<div class="icon"><i class="far fa-bookmark" aria-hidden="true"></i></div>\n            <h4>Catalogue</h4>\n            <p>Browse the list or use the dropdown filter to find an application that supports a specific activity.</p>',s=u("\n        "),o=c("div"),m=c("select"),g=c("option"),y=u("Find an app to..."),b=c("option"),x=u("...create learning resources"),C=c("option"),N=u("...communicate with students electronically"),k=c("option"),w=u("...act as a platform for collaboration"),S=c("option"),O=u("...collect, organise and share content"),j=c("option"),E=u("...record an event or artefact"),L=c("option"),P=u("...create interactive learning opportunities"),T=c("option"),$=u("...gauge students understanding"),B=c("option"),M=u("All apps"),A=u("\n        "),F=c("div"),H=c("div"),q=c("ul");for(var t=0;t<G.length;t+=1)G[t].c();this.h()},h:function(){r.className="bs-callout bs-callout-catalogue clearfix",g.selected=!0,g.__value="all",g.value=g.__value,b.__value="create",b.value=b.__value,C.__value="connect",C.value=C.__value,k.__value="collaborate",k.value=k.__value,S.__value="curate",S.value=S.__value,j.__value="capture",j.value=j.__value,L.__value="captivate",L.value=L.__value,T.__value="check",T.value=T.__value,B.__value="all",B.value=B.__value,m.id="filterText",m.className="btn btn-secondary dropdown-toggle btn-block",d(m,"change",z),o.className="dropdown",q.className="list-group list-group-flush",H.className="front",F.className="card cataloguepane",F.dataset.spy="scroll",n.className="col-md-12",i.className="card-container manual-flip"},m:function(t,e){f(i,t,e),_(n,i),_(r,n),_(s,n),_(o,n),_(m,o),_(g,m),_(y,g),_(b,m),_(x,b),_(C,m),_(N,C),_(k,m),_(w,k),_(S,m),_(O,S),_(j,m),_(E,j),_(L,m),_(P,L),_(T,m),_($,T),_(B,m),_(M,B),_(A,n),_(F,n),_(H,F),_(q,H);for(var a=0;a<G.length;a+=1)G[a].m(q,null)},p:function(e,i){var n=i.filteredCards;if(e.filteredCards){for(var r=0;r<n.length;r+=1){var s=l({},i,{card:n[r],card_index:r});G[r]?G[r].p(e,s):(G[r]=a(t,s),G[r].c(),G[r].m(q,null))}for(;r<G.length;r+=1)G[r].u(),G[r].d();G.length=n.length}},u:function(){h(i);for(var t=0;t<G.length;t+=1)G[t].u()},d:function(){v(m,"change",z),p(G)}}}function a(t,e){for(var i,n,a,o,g,y,b,x,C,N,k,w,S,O,j,E,L,P,T,$=e.card,B=(e.card_index,$.name),M=$.submitter.name,A=$.submitter.title,F=$.tagline,H=$.activities,q=[],z=0;z<H.length;z+=1)q[z]=r(t,l({},e,{activity:H[z],activity_index:z}));return{c:function(){i=c("a"),n=c("div"),a=c("h5"),o=c("img"),y=u(" "),b=u(B),x=u("\n                            "),C=c("p");for(var t=0;t<q.length;t+=1)q[t].c();N=u("\n                            "),k=c("p"),w=c("small"),S=u("Submitted by "),O=u(M),j=u(" | "),E=u(A),L=u("\n                            "),P=c("p"),T=u(F),this.h()},h:function(){m(o,"display","inline"),m(o,"height","1em"),m(o,"vertical-align","top"),o.src=g=$.assets.logo.image,C.className="categories",k.className="submittercatalogue",P.className="appinfo",n.className="applist",i.href="#",i.className="list-group-item list-group-item-action",d(i,"click",s),i._svelte={component:t,filteredCards:e.filteredCards,card_index:e.card_index}},m:function(t,e){f(i,t,e),_(n,i),_(a,n),_(o,a),_(y,a),_(b,a),_(x,n),_(C,n);for(var r=0;r<q.length;r+=1)q[r].m(C,null);_(N,n),_(k,n),_(w,k),_(S,w),_(O,w),_(j,w),_(E,w),_(L,n),_(P,n),_(T,P)},p:function(e,n){$=n.card,n.card_index,e.filteredCards&&g!==(g=$.assets.logo.image)&&(o.src=g),e.filteredCards&&B!==(B=$.name)&&(b.data=B);var a=$.activities;if(e.filteredCards){for(var s=0;s<a.length;s+=1){var c=l({},n,{activity:a[s],activity_index:s});q[s]?q[s].p(e,c):(q[s]=r(t,c),q[s].c(),q[s].m(C,null))}for(;s<q.length;s+=1)q[s].u(),q[s].d();q.length=a.length}e.filteredCards&&M!==(M=$.submitter.name)&&(O.data=M),e.filteredCards&&A!==(A=$.submitter.title)&&(E.data=A),e.filteredCards&&F!==(F=$.tagline)&&(T.data=F),i._svelte.filteredCards=n.filteredCards,i._svelte.card_index=n.card_index},u:function(){h(i);for(var t=0;t<q.length;t+=1)q[t].u()},d:function(){p(q),v(i,"click",s)}}}function r(t,e){e.card;var i,n,a,r,s=e.activity,o=(e.card_index,e.activity_index,s.name);return{c:function(){i=c("span"),n=u(o),r=u(" "),this.h()},h:function(){i.className=a="badge activity "+s.name},m:function(t,e){f(i,t,e),_(n,i),f(r,t,e)},p:function(t,e){e.card,s=e.activity,e.card_index,e.activity_index,t.filteredCards&&o!==(o=s.name)&&(n.data=o),t.filteredCards&&a!==(a="badge activity "+s.name)&&(i.className=a)},u:function(){h(i),h(r)},d:g}}function s(t){var e=this._svelte.component,i=this._svelte.filteredCards[this._svelte.card_index];e.store.set({currentPage:"card",currentCard:i.id})}function o(t){y(this,t),this._state=l(this.store._init(["cards"]),i(),t.data),this.store._add(this,["cards"]),this._recompute({$cards:1,filter:1},this._state),this._handlers.destroy=[b],this._fragment=n(this,this._state),t.target&&(this._fragment.c(),this._fragment.m(t.target,t.anchor||null))}function l(t){for(var e,i,n=1,a=arguments.length;n<a;n++)for(e in i=arguments[n])t[e]=i[e];return t}function c(t){return document.createElement(t)}function u(t){return document.createTextNode(t)}function d(t,e,i){t.addEventListener(e,i,!1)}function f(t,e,i){e.insertBefore(t,i)}function _(t,e){e.appendChild(t)}function h(t){t.parentNode.removeChild(t)}function v(t,e,i){t.removeEventListener(e,i,!1)}function p(t){for(var e=0;e<t.length;e+=1)t[e]&&t[e].d()}function m(t,e,i){t.style.setProperty(e,i)}function g(){}function y(t,e){t._observers={pre:P(),post:P()},t._handlers=P(),t._bind=e._bind,t.options=e,t.root=e.root||t,t.store=t.root.store||e.store}function b(){this.store._remove(this)}function x(t){this.destroy=g,this.fire("destroy"),this.set=this.get=g,!1!==t&&this._fragment.u(),this._fragment.d(),this._fragment=this._state=null}function C(t){return t?this._state[t]:this._state}function N(t,e){var i=t in this._handlers&&this._handlers[t].slice();if(i)for(var n=0;n<i.length;n+=1)i[n].call(this,e)}function k(t,e,i){var n=i&&i.defer?this._observers.post:this._observers.pre;return(n[t]||(n[t]=[])).push(e),i&&!1===i.init||(e.__calling=!0,e.call(this,this._state[t]),e.__calling=!1),{cancel:function(){var i=n[t].indexOf(e);~i&&n[t].splice(i,1)}}}function w(t,e){if("teardown"===t)return this.on("destroy",e);var i=this._handlers[t]||(this._handlers[t]=[]);return i.push(e),{cancel:function(){var t=i.indexOf(e);~t&&i.splice(t,1)}}}function S(t){this._set(l({},t)),this.root._lock||(this.root._lock=!0,T(this.root._beforecreate),T(this.root._oncreate),T(this.root._aftercreate),this.root._lock=!1)}function O(t){var e=this._state,i={},n=!1;for(var a in t)this._differs(t[a],e[a])&&(i[a]=n=!0);n&&(this._state=l({},e,t),this._recompute(i,this._state),this._bind&&this._bind(i,this._state),this._fragment&&($(this,this._observers.pre,i,this._state,e),this._fragment.p(i,this._state),$(this,this._observers.post,i,this._state,e)))}function j(t,e){this._fragment.m(t,e)}function E(){this._fragment&&this._fragment.u()}function L(e,i){return e!=e?i==i:e!==i||e&&"object"===(void 0===e?"undefined":t(e))||"function"==typeof e}function P(){return Object.create(null)}function T(t){for(;t&&t.length;)t.shift()()}function $(t,e,i,n,a){for(var r in e)if(i[r]){var s=n[r],o=a[r],l=e[r];if(l)for(var c=0;c<l.length;c+=1){var u=l[c];u.__calling||(u.__calling=!0,u.call(t,s,o),u.__calling=!1)}}}l(o.prototype,{destroy:x,get:C,fire:N,observe:k,on:w,set:S,teardown:x,_set:O,_mount:j,_unmount:E,_differs:L}),o.prototype._recompute=function(t,i){(t.$cards||t.filter)&&this._differs(i.filteredCards,i.filteredCards=e(i.$cards,i.filter))&&(t.filteredCards=!0)},exports.default=o;
+},{}],34:[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0});var t="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t},e={resetForm:function(){document.getElementById("gform").style.display="block",document.getElementById("thankyou_message").style.display="none",document.getElementById("gform").reset()},submitForm:function(t){!function(t){t.preventDefault();var e=function(){var t=document.getElementById("gform").elements,e=Object.keys(t).map(function(e){return void 0!==t[e].name?t[e].name:t[e].length>0?t[e].item(0).name:void 0}).filter(function(t,e,n){return n.indexOf(t)==e&&t}),n={};return e.forEach(function(e){n[e]=t[e].value;var o="";if("checkbox"===t[e].type)o=o+t[e].checked+", ",n[e]=o.slice(0,-2);else if(t[e].length)for(var i=0;i<t[e].length;i++)t[e].item(i).checked&&(o=o+t[e].item(i).value+", ",n[e]=o.slice(0,-2))}),n}();if(e.honeypot)return!1;if(o=e.email,!/^([\w-]+(?:\.[\w-]+)*)@((?:[\w-]+\.)*\w[\w-]{0,66})\.([a-z]{2,6}(?:\.[a-z]{2})?)$/i.test(o))return document.getElementById("Email").style.display="block",!1;var n=new XMLHttpRequest;n.open("POST","https://script.google.com/macros/s/AKfycbwdPK1Xn_XdIKYYnMLltp6u51nGpuRltLkwPvMUyLmMY1biPhpz/exec"),n.setRequestHeader("Content-Type","application/x-www-form-urlencoded"),n.onreadystatechange=function(){document.getElementById("gform").style.display="none",document.getElementById("thankyou_message").style.display="block"};var o,i=Object.keys(e).map(function(t){return encodeURIComponent(t)+"="+encodeURIComponent(e[t])}).join("&");n.send(i)}(t)}};function n(t,e){var n,o,m,d,h,p,b,_,g,y,v,k,w;function x(e){t.resetForm()}function E(e){t.submitForm(e)}return{c:function(){n=i("div"),o=i("div"),(m=i("div")).innerHTML='<div class="icon"><i class="fa fa-plus-square-o" aria-hidden="true"></i></div>\n            <h4>Contribute</h4>\n            <p>Help others discover new technologies by telling us about the apps you use.</p>\n            <p class="instruction">Complete the form below and we\'ll be in touch to discuss creating a custom card for your suggestion.</p>',d=s("\n        "),h=i("div"),(p=i("div")).innerHTML='<i class="fa fa-thumbs-o-up" aria-hidden="true"></i>',b=s("\n            "),_=i("p"),(g=i("strong")).textContent="Alright! ",y=s("Thanks for making a suggestion. "),(v=i("a")).textContent="Click here to make another.",k=s("\n        "),(w=i("form")).innerHTML='<div class="form-group"><label for="name">Name:</label>\n                <input type="text" class="form-control" id="name" name="name" aria-describedby="emailHelp" placeholder=""></div>\n            <div class="form-group"><label for="email">Email address:</label>\n                <input type="email" class="form-control" id="email" name="email" aria-describedby="emailHelp" placeholder=""></div>\n            <div class="form-group"><label for="location">School/Faculty/Directorate:</label>\n            <input class="form-control" id="location" name="location" aria-describedby="emailHelp" placeholder=""></div>\n            <div class="form-group"><label for="message">App suggestion:</label>\n                <textarea class="form-control" id="message" name="message" rows="3"></textarea></div>\n            <div class="form-group"><label class="sr-only">Keep this field blank</label>\n                <input id="honeypot" type="text" name="honeypot" value=""></div>\n            <button type="submit" class="btn btn-block submitbutton">Submit</button>',this.h()},h:function(){m.className="bs-callout bs-callout-contribute clearfix",p.className="icon",v.href="#",r(v,"click",x),h.className="bs-callout bs-thanks clearfix",h.id="thankyou_message",w.id="gform",r(w,"submit",E),o.className="col-md-12"},m:function(t,e){a(n,t,e),l(o,n),l(m,o),l(d,o),l(h,o),l(p,h),l(b,h),l(_,h),l(g,_),l(y,_),l(v,_),l(k,o),l(w,o)},p:c,u:function(){u(n)},d:function(){f(v,"click",x),f(w,"submit",E)}}}function o(t){m(this,t),this._state=d({},t.data),this._fragment=n(this,this._state),t.target&&(this._fragment.c(),this._fragment.m(t.target,t.anchor||null))}function i(t){return document.createElement(t)}function s(t){return document.createTextNode(t)}function r(t,e,n){t.addEventListener(e,n,!1)}function a(t,e,n){e.insertBefore(t,n)}function l(t,e){e.appendChild(t)}function c(){}function u(t){t.parentNode.removeChild(t)}function f(t,e,n){t.removeEventListener(e,n,!1)}function m(t,e){t._observers={pre:E(),post:E()},t._handlers=E(),t._bind=e._bind,t.options=e,t.root=e.root||t,t.store=t.root.store||e.store}function d(t){for(var e,n,o=1,i=arguments.length;o<i;o++)for(e in n=arguments[o])t[e]=n[e];return t}function h(t){this.destroy=c,this.fire("destroy"),this.set=this.get=c,!1!==t&&this._fragment.u(),this._fragment.d(),this._fragment=this._state=null}function p(t){return t?this._state[t]:this._state}function b(t,e){var n=t in this._handlers&&this._handlers[t].slice();if(n)for(var o=0;o<n.length;o+=1)n[o].call(this,e)}function _(t,e,n){var o=n&&n.defer?this._observers.post:this._observers.pre;return(o[t]||(o[t]=[])).push(e),n&&!1===n.init||(e.__calling=!0,e.call(this,this._state[t]),e.__calling=!1),{cancel:function(){var n=o[t].indexOf(e);~n&&o[t].splice(n,1)}}}function g(t,e){if("teardown"===t)return this.on("destroy",e);var n=this._handlers[t]||(this._handlers[t]=[]);return n.push(e),{cancel:function(){var t=n.indexOf(e);~t&&n.splice(t,1)}}}function y(t){this._set(d({},t)),this.root._lock||(this.root._lock=!0,C(this.root._beforecreate),C(this.root._oncreate),C(this.root._aftercreate),this.root._lock=!1)}function v(t){var e=this._state,n={},o=!1;for(var i in t)this._differs(t[i],e[i])&&(n[i]=o=!0);o&&(this._state=d({},e,t),this._recompute(n,this._state),this._bind&&this._bind(n,this._state),this._fragment&&(I(this,this._observers.pre,n,this._state,e),this._fragment.p(n,this._state),I(this,this._observers.post,n,this._state,e)))}function k(t,e){this._fragment.m(t,e)}function w(){this._fragment&&this._fragment.u()}function x(e,n){return e!=e?n==n:e!==n||e&&"object"===(void 0===e?"undefined":t(e))||"function"==typeof e}function E(){return Object.create(null)}function C(t){for(;t&&t.length;)t.shift()()}function I(t,e,n,o,i){for(var s in e)if(n[s]){var r=o[s],a=i[s],l=e[s];if(l)for(var c=0;c<l.length;c+=1){var u=l[c];u.__calling||(u.__calling=!0,u.call(t,r,a),u.__calling=!1)}}}d(o.prototype,e,{destroy:h,get:p,fire:b,observe:_,on:g,set:y,teardown:h,_set:v,_mount:k,_unmount:w,_differs:x}),o.prototype._recompute=c,exports.default=o;
+},{}],35:[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0});var t="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t};function e(t,e){var r;return{c:function(){(r=n("div")).innerHTML='<div class="col-md-12"><div class="bs-callout bs-callout-about clearfix"><div class="icon"><i class="fa fa-info" aria-hidden="true"></i></div>\n\n            <h4>About</h4>\n            <p>This work is licensed under a <a rel="license" href="http://creativecommons.org/licenses/by-nc/4.0/" target="_blank">Creative Commons Attribution-NonCommercial 4.0 International License</a>.</p>\n            <p>All resources are freely available via <a href="https://github.com/danieladamharding/tel_cards" target="_blank">this GitHub repository</a>.</p>\n            <p>Produced by <a href="https://www.twitter.com/danielharding" target="_blank">Dan Harding</a>, <a href="https://www.twitter.com/Matthew_Street" target="_blank">Matthew Street</a> and <a href="https://www.twitter.com/confusedmatrix" target="_blank">Chris Briggs</a> at <a href="https://www.keele.ac.uk" target="_blank">Keele University</a>, UK.</p>\n            <a rel="license" href="http://creativecommons.org/licenses/by-nc/4.0/" target="_blank"><img alt="Creative Commons License" class="creativecommons" style="border-width: 0" src="https://i.creativecommons.org/l/by-nc/4.0/88x31.png"></a><br></div></div>'},m:function(t,e){i(r,t,e)},p:s,u:function(){o(r)},d:s}}function r(t){a(this,t),this._state=c({},t.data),this._fragment=e(this,this._state),t.target&&(this._fragment.c(),this._fragment.m(t.target,t.anchor||null))}function n(t){return document.createElement(t)}function i(t,e,r){e.insertBefore(t,r)}function s(){}function o(t){t.parentNode.removeChild(t)}function a(t,e){t._observers={pre:v(),post:v()},t._handlers=v(),t._bind=e._bind,t.options=e,t.root=e.root||t,t.store=t.root.store||e.store}function c(t){for(var e,r,n=1,i=arguments.length;n<i;n++)for(e in r=arguments[n])t[e]=r[e];return t}function h(t){this.destroy=s,this.fire("destroy"),this.set=this.get=s,!1!==t&&this._fragment.u(),this._fragment.d(),this._fragment=this._state=null}function l(t){return t?this._state[t]:this._state}function f(t,e){var r=t in this._handlers&&this._handlers[t].slice();if(r)for(var n=0;n<r.length;n+=1)r[n].call(this,e)}function u(t,e,r){var n=r&&r.defer?this._observers.post:this._observers.pre;return(n[t]||(n[t]=[])).push(e),r&&!1===r.init||(e.__calling=!0,e.call(this,this._state[t]),e.__calling=!1),{cancel:function(){var r=n[t].indexOf(e);~r&&n[t].splice(r,1)}}}function _(t,e){if("teardown"===t)return this.on("destroy",e);var r=this._handlers[t]||(this._handlers[t]=[]);return r.push(e),{cancel:function(){var t=r.indexOf(e);~t&&r.splice(t,1)}}}function d(t){this._set(c({},t)),this.root._lock||(this.root._lock=!0,y(this.root._beforecreate),y(this.root._oncreate),y(this.root._aftercreate),this.root._lock=!1)}function m(t){var e=this._state,r={},n=!1;for(var i in t)this._differs(t[i],e[i])&&(r[i]=n=!0);n&&(this._state=c({},e,t),this._recompute(r,this._state),this._bind&&this._bind(r,this._state),this._fragment&&(w(this,this._observers.pre,r,this._state,e),this._fragment.p(r,this._state),w(this,this._observers.post,r,this._state,e)))}function p(t,e){this._fragment.m(t,e)}function b(){this._fragment&&this._fragment.u()}function g(e,r){return e!=e?r==r:e!==r||e&&"object"===(void 0===e?"undefined":t(e))||"function"==typeof e}function v(){return Object.create(null)}function y(t){for(;t&&t.length;)t.shift()()}function w(t,e,r,n,i){for(var s in e)if(r[s]){var o=n[s],a=i[s],c=e[s];if(c)for(var h=0;h<c.length;h+=1){var l=c[h];l.__calling||(l.__calling=!0,l.call(t,o,a),l.__calling=!1)}}}c(r.prototype,{destroy:h,get:l,fire:f,observe:u,on:_,set:d,teardown:h,_set:m,_mount:p,_unmount:b,_differs:g}),r.prototype._recompute=s,exports.default=r;
+},{}],7:[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0});var t="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t},e=require("./Card.svelte"),a=u(e),r=require("./Catalogue.svelte"),n=u(r),i=require("./Contribute.svelte"),s=u(i),o=require("./About.svelte"),c=u(o);function u(t){return t&&t.__esModule?t:{default:t}}function l(t,e){return t.find(function(t){return t.id===e})}function f(){}function d(t,e){var r,i,o,u,l,f,d,h,$,N,k,w,C,L,M,T,q,x,S,j,H,O,E,A,B,z,D,F,G,I;function J(e){t.store.set({currentPage:"card"})}function K(e){t.store.set({currentPage:"catalogue"})}function Q(e){t.store.set({currentPage:"contribute"})}function R(e){t.store.set({currentPage:"about"})}var U=new a.default({root:t.root,data:{card:e.card}}),V=new n.default({root:t.root}),W=new s.default({root:t.root}),X=new c.default({root:t.root});return{c:function(){r=_("div"),i=_("div"),o=_("div"),u=_("ul"),l=_("li"),(f=_("a")).innerHTML='<i class="fas fa-magic" aria-hidden="true"></i>',h=_("li"),($=_("a")).innerHTML='<i class="far fa-bookmark" aria-hidden="true"></i>',k=_("li"),(w=_("a")).innerHTML='<i class="far fa-plus-square" aria-hidden="true"></i>',L=_("li"),(M=_("a")).innerHTML='<i class="fas fa-info" aria-hidden="true"></i>',q=b("\n\t\t"),x=_("div"),S=_("div"),j=_("div"),U._fragment.c(),O=b("\n\t\t\t"),E=_("div"),V._fragment.c(),B=b("\n\t\t\t"),z=_("div"),W._fragment.c(),F=b("\n\t\t\t"),G=_("div"),X._fragment.c(),this.h()},h:function(){f.className=d="nav-link "+("card"===e.$currentPage?"active":""),f.id="app-tab",f.href="#appgenerator",m(f,"click",J),l.className="nav-item",$.className=N="nav-link "+("catalogue"===e.$currentPage?"active show":""),$.id="catalogue-tab",$.href="#catalogue",m($,"click",K),h.className="nav-item",w.className=C="nav-link "+("contribute"===e.$currentPage?"active show":""),w.id="contribute-tab",w.href="#contribute",m(w,"click",Q),k.className="nav-item",M.className=T="nav-link "+("about"===e.$currentPage?"active show":""),M.id="about-tab",M.href="#about",m(M,"click",R),L.className="nav-item",u.className="nav nav-tabs card-header-tabs justify-content-center",o.className="card-header",v(o,"role","navigation"),j.className="container",S.className=H="tab-pane fade "+("card"===e.$currentPage?"active show":""),S.id="appgenerator",v(S,"role","tabpanel"),v(S,"aria-labelledby","app-tab"),E.className=A="tab-pane fade "+("catalogue"===e.$currentPage?"active show":""),E.id="catalogue",v(E,"role","tabpanel"),v(E,"aria-labelledby","catalogue-tab"),z.className=D="tab-pane fade "+("contribute"===e.$currentPage?"active show":""),z.id="contribute",v(z,"role","tabpanel"),v(z,"aria-labelledby","contribute-tab"),G.className=I="tab-pane fade "+("about"===e.$currentPage?"active show":""),G.id="about",v(G,"role","tabpanel"),v(G,"aria-labelledby","about-tab"),x.className="tab-content",x.id="myTabContent",i.className="card mainCard",r.className="col-md-12"},m:function(t,e){g(r,t,e),p(i,r),p(o,i),p(u,o),p(l,u),p(f,l),p(h,u),p($,h),p(k,u),p(w,k),p(L,u),p(M,L),p(q,i),p(x,i),p(S,x),p(j,S),U._mount(j,null),p(O,x),p(E,x),V._mount(E,null),p(B,x),p(z,x),W._mount(z,null),p(F,x),p(G,x),X._mount(G,null)},p:function(t,e){t.$currentPage&&d!==(d="nav-link "+("card"===e.$currentPage?"active":""))&&(f.className=d),t.$currentPage&&N!==(N="nav-link "+("catalogue"===e.$currentPage?"active show":""))&&($.className=N),t.$currentPage&&C!==(C="nav-link "+("contribute"===e.$currentPage?"active show":""))&&(w.className=C),t.$currentPage&&T!==(T="nav-link "+("about"===e.$currentPage?"active show":""))&&(M.className=T);var a={};t.card&&(a.card=e.card),U._set(a),t.$currentPage&&H!==(H="tab-pane fade "+("card"===e.$currentPage?"active show":""))&&(S.className=H),t.$currentPage&&A!==(A="tab-pane fade "+("catalogue"===e.$currentPage?"active show":""))&&(E.className=A),t.$currentPage&&D!==(D="tab-pane fade "+("contribute"===e.$currentPage?"active show":""))&&(z.className=D),t.$currentPage&&I!==(I="tab-pane fade "+("about"===e.$currentPage?"active show":""))&&(G.className=I)},u:function(){y(r)},d:function(){P(f,"click",J),P($,"click",K),P(w,"click",Q),P(M,"click",R),U.destroy(!1),V.destroy(!1),W.destroy(!1),X.destroy(!1)}}}function h(t){$(this,t),this._state=N(this.store._init(["cards","currentCard","currentPage"]),t.data),this.store._add(this,["cards","currentCard","currentPage"]),this._recompute({$cards:1,$currentCard:1},this._state),this._handlers.destroy=[k];var e=f.bind(this);t.root||(this._oncreate=[],this._beforecreate=[],this._aftercreate=[]),this._fragment=d(this,this._state),this.root._oncreate.push(e),t.target&&(this._fragment.c(),this._fragment.m(t.target,t.anchor||null),this._lock=!0,w(this._beforecreate),w(this._oncreate),w(this._aftercreate),this._lock=!1)}function _(t){return document.createElement(t)}function b(t){return document.createTextNode(t)}function m(t,e,a){t.addEventListener(e,a,!1)}function v(t,e,a){t.setAttribute(e,a)}function g(t,e,a){e.insertBefore(t,a)}function p(t,e){e.appendChild(t)}function y(t){t.parentNode.removeChild(t)}function P(t,e,a){t.removeEventListener(e,a,!1)}function $(t,e){t._observers={pre:E(),post:E()},t._handlers=E(),t._bind=e._bind,t.options=e,t.root=e.root||t,t.store=t.root.store||e.store}function N(t){for(var e,a,r=1,n=arguments.length;r<n;r++)for(e in a=arguments[r])t[e]=a[e];return t}function k(){this.store._remove(this)}function w(t){for(;t&&t.length;)t.shift()()}function C(t){this.destroy=A,this.fire("destroy"),this.set=this.get=A,!1!==t&&this._fragment.u(),this._fragment.d(),this._fragment=this._state=null}function L(t){return t?this._state[t]:this._state}function M(t,e){var a=t in this._handlers&&this._handlers[t].slice();if(a)for(var r=0;r<a.length;r+=1)a[r].call(this,e)}function T(t,e,a){var r=a&&a.defer?this._observers.post:this._observers.pre;return(r[t]||(r[t]=[])).push(e),a&&!1===a.init||(e.__calling=!0,e.call(this,this._state[t]),e.__calling=!1),{cancel:function(){var a=r[t].indexOf(e);~a&&r[t].splice(a,1)}}}function q(t,e){if("teardown"===t)return this.on("destroy",e);var a=this._handlers[t]||(this._handlers[t]=[]);return a.push(e),{cancel:function(){var t=a.indexOf(e);~t&&a.splice(t,1)}}}function x(t){this._set(N({},t)),this.root._lock||(this.root._lock=!0,w(this.root._beforecreate),w(this.root._oncreate),w(this.root._aftercreate),this.root._lock=!1)}function S(t){var e=this._state,a={},r=!1;for(var n in t)this._differs(t[n],e[n])&&(a[n]=r=!0);r&&(this._state=N({},e,t),this._recompute(a,this._state),this._bind&&this._bind(a,this._state),this._fragment&&(B(this,this._observers.pre,a,this._state,e),this._fragment.p(a,this._state),B(this,this._observers.post,a,this._state,e)))}function j(t,e){this._fragment.m(t,e)}function H(){this._fragment&&this._fragment.u()}function O(e,a){return e!=e?a==a:e!==a||e&&"object"===(void 0===e?"undefined":t(e))||"function"==typeof e}function E(){return Object.create(null)}function A(){}function B(t,e,a,r,n){for(var i in e)if(a[i]){var s=r[i],o=n[i],c=e[i];if(c)for(var u=0;u<c.length;u+=1){var l=c[u];l.__calling||(l.__calling=!0,l.call(t,s,o),l.__calling=!1)}}}N(h.prototype,{destroy:C,get:L,fire:M,observe:T,on:q,set:x,teardown:C,_set:S,_mount:j,_unmount:H,_differs:O}),h.prototype._recompute=function(t,e){(t.$cards||t.$currentCard)&&this._differs(e.card,e.card=l(e.$cards,e.$currentCard))&&(t.card=!0)},exports.default=h;
+},{"./Card.svelte":32,"./Catalogue.svelte":33,"./Contribute.svelte":34,"./About.svelte":35}],36:[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0});var t="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t};function e(t,e){var a;return{c:function(){(a=i("div")).innerHTML='<div class="col-md-12"><div class="bs-callout bs-callout-support clearfix"><div class="icon"><i class="fa fa-comments-o" aria-hidden="true"></i></div>\n            <h4>Support</h4>\n            <p>If you have any questions or issues relating to the apps featured here, please speak to your local learning technology support person.</p> <p>A list of University contacts can be found below.</p></div>\n        <div class="contacts-nav"><ul class="nav nav-pills nav-fill" role="tablist"><li class="nav-item"><a class="nav-link active" href="#humss" role="tab" data-toggle="pill">Faculty of HumSS</a></li>\n                <li class="nav-item"><a class="nav-link" href="#health" role="tab" data-toggle="pill">Faculty of Medicine and Health Sciences</a></li>\n                <li class="nav-item"><a class="nav-link" href="#naturalsciences" role="tab" data-toggle="pill">Faculty of Natural Sciences</a></li>\n                <li class="nav-item"><a class="nav-link" href="#lpdc" role="tab" data-toggle="pill">LPDC</a></li>\n            </ul>\n            <div class="tab-content"><div role="tabpanel" class="tab-pane fade show active" id="humss"><br>\n                    <div class="list-group"><a href="mailto:d.harding@keele.ac.uk?Subject=KLE%20Templates" class="list-group-item list-group-item-action"><i class="fa fa-envelope" aria-hidden="true"></i>   Dan Harding</a>\n                        <a href="mailto:c.hallam@keele.ac.uk?Subject=KLE%20Templates" class="list-group-item list-group-item-action"><i class="fa fa-envelope" aria-hidden="true"></i>   Cat Hallam</a>\n                        <a href="mailto:a.gater@keele.ac.uk?Subject=KLE%20Templates" class="list-group-item list-group-item-action"><i class="fa fa-envelope" aria-hidden="true"></i>   Andy Gater (Language Centre)</a></div></div>\n                <div role="tabpanel" class="tab-pane fade" id="health"><br>\n                    <div class="list-group"><a href="mailto:t.smale@keele.ac.uk?Subject=KLE%20Templates" class="list-group-item list-group-item-action"><i class="fa fa-envelope" aria-hidden="true"></i>   Tim Smale</a>\n                        <a href="mailto:p.r.lonsdale@keele.ac.uk?Subject=KLE%20Templates" class="list-group-item list-group-item-action"><i class="fa fa-envelope" aria-hidden="true"></i>   Pete Lonsdale (School of Nursing and Midwifery)</a></div></div>\n                <div role="tabpanel" class="tab-pane fade" id="naturalsciences"><br>\n                    <div class="list-group"><a href="mailto:d.a.kerr@keele.ac.uk?Subject=KLE%20Templates" class="list-group-item list-group-item-action"><i class="fa fa-envelope" aria-hidden="true"></i>   Daryl Kerr</a></div></div>\n                <div role="tabpanel" class="tab-pane fade" id="lpdc"><br>    \n                    <div class="list-group"><a href="mailto:m.j.street@keele.ac.uk?Subject=KLE%20Templates" class="list-group-item list-group-item-action"><i class="fa fa-envelope" aria-hidden="true"></i>   Matthew Street</a></div></div></div></div></div>'},m:function(t,e){s(a,t,e)},p:n,u:function(){l(a)},d:n}}function a(t){r(this,t),this._state=o({},t.data),this._fragment=e(this,this._state),t.target&&(this._fragment.c(),this._fragment.m(t.target,t.anchor||null))}function i(t){return document.createElement(t)}function s(t,e,a){e.insertBefore(t,a)}function n(){}function l(t){t.parentNode.removeChild(t)}function r(t,e){t._observers={pre:b(),post:b()},t._handlers=b(),t._bind=e._bind,t.options=e,t.root=e.root||t,t.store=t.root.store||e.store}function o(t){for(var e,a,i=1,s=arguments.length;i<s;i++)for(e in a=arguments[i])t[e]=a[e];return t}function c(t){this.destroy=n,this.fire("destroy"),this.set=this.get=n,!1!==t&&this._fragment.u(),this._fragment.d(),this._fragment=this._state=null}function u(t){return t?this._state[t]:this._state}function f(t,e){var a=t in this._handlers&&this._handlers[t].slice();if(a)for(var i=0;i<a.length;i+=1)a[i].call(this,e)}function h(t,e,a){var i=a&&a.defer?this._observers.post:this._observers.pre;return(i[t]||(i[t]=[])).push(e),a&&!1===a.init||(e.__calling=!0,e.call(this,this._state[t]),e.__calling=!1),{cancel:function(){var a=i[t].indexOf(e);~a&&i[t].splice(a,1)}}}function d(t,e){if("teardown"===t)return this.on("destroy",e);var a=this._handlers[t]||(this._handlers[t]=[]);return a.push(e),{cancel:function(){var t=a.indexOf(e);~t&&a.splice(t,1)}}}function p(t){this._set(o({},t)),this.root._lock||(this.root._lock=!0,y(this.root._beforecreate),y(this.root._oncreate),y(this.root._aftercreate),this.root._lock=!1)}function v(t){var e=this._state,a={},i=!1;for(var s in t)this._differs(t[s],e[s])&&(a[s]=i=!0);i&&(this._state=o({},e,t),this._recompute(a,this._state),this._bind&&this._bind(a,this._state),this._fragment&&(k(this,this._observers.pre,a,this._state,e),this._fragment.p(a,this._state),k(this,this._observers.post,a,this._state,e)))}function m(t,e){this._fragment.m(t,e)}function _(){this._fragment&&this._fragment.u()}function g(e,a){return e!=e?a==a:e!==a||e&&"object"===(void 0===e?"undefined":t(e))||"function"==typeof e}function b(){return Object.create(null)}function y(t){for(;t&&t.length;)t.shift()()}function k(t,e,a,i,s){for(var n in e)if(a[n]){var l=i[n],r=s[n],o=e[n];if(o)for(var c=0;c<o.length;c+=1){var u=o[c];u.__calling||(u.__calling=!0,u.call(t,l,r),u.__calling=!1)}}}o(a.prototype,{destroy:c,get:u,fire:f,observe:h,on:d,set:p,teardown:c,_set:v,_mount:m,_unmount:_,_differs:g}),a.prototype._recompute=n,exports.default=a;
+},{}],8:[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0});var t="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t},e=require("./Card.svelte"),a=c(e),n=require("./Contribute.svelte"),r=c(n),i=require("./Support.svelte"),o=c(i),s=require("./About.svelte"),l=c(s);function c(t){return t&&t.__esModule?t:{default:t}}function u(t,e){return console.log(e),"all"===e?t:t.filter(function(t){return-1!==t.activities.map(function(t){return t.name}).indexOf(e)})}function f(){return{filter:"all"}}function d(t,e){var a,n,i,s,c,u,f,d,h,k,S,q,A,O,j,E,L,M,T,$,B,F,H,P,z,D,G,I,J,K,Q,R,U,V,W,X;function Y(e){t.set({filter:e.target.value})}for(var Z=e.filteredCards,tt=[],et=0;et<Z.length;et+=1)tt[et]=_(t,v({},e,{card:Z[et],card_index:et}));var at=new r.default({root:t.root}),nt=new l.default({root:t.root}),rt=new o.default({root:t.root});return{c:function(){a=p("div"),n=p("div"),i=p("div"),s=p("select"),c=p("option"),u=b("Find an app to..."),f=p("option"),d=b("...create learning resources"),h=p("option"),k=b("...communicate with students electronically"),S=p("option"),q=b("...act as a platform for collaboration"),A=p("option"),O=b("...collect, organise and share content"),j=p("option"),E=b("...record an event or artefact"),L=p("option"),M=b("...create interactive learning opportunities"),T=p("option"),$=b("...gauge students understanding"),B=p("option"),F=b("All apps"),H=b("\n\n    "),P=p("div"),z=p("div");for(var t=0;t<tt.length;t+=1)tt[t].c();D=b("\n\n    "),G=p("div"),I=p("div"),(J=p("div")).innerHTML='<ul class="nav nav-pills nav-fill" id="pills-tab" role="tablist"><li class="nav-item"><a class="nav-link active" id="contribute-tab" data-toggle="pill" href="#contribute" role="tab" aria-controls="Contribute" aria-selected="true"><i class="fa fa-plus-square-o" aria-hidden="true"></i>  Contribute</a></li>\n                    <li class="nav-item"><a class="nav-link" id="support-tab" data-toggle="pill" href="#support" role="tab" aria-controls="Support" aria-selected="false"><i class="fa fa-comments-o" aria-hidden="true"></i>  Support</a></li>\n                    <li class="nav-item"><a class="nav-link" id="about-tab" data-toggle="pill" href="#about" role="tab" aria-controls="About" aria-selected="false"><i class="fa fa-info" aria-hidden="true"></i>  About</a></li>\n                </ul>',K=b("\n\n            "),Q=p("div"),R=p("div"),at._fragment.c(),U=b("\n                "),V=p("div"),nt._fragment.c(),W=b("\n                "),X=p("div"),rt._fragment.c(),this.h()},h:function(){c.selected=!0,c.__value="all",c.value=c.__value,f.__value="create",f.value=f.__value,h.__value="connect",h.value=h.__value,S.__value="collaborate",S.value=S.__value,A.__value="curate",A.value=A.__value,j.__value="capture",j.value=j.__value,L.__value="captivate",L.value=L.__value,T.__value="check",T.value=T.__value,B.__value="all",B.value=B.__value,s.id="filterText",s.className="btn btn-secondary dropdown-toggle btn-lg btn-block",m(s,"change",Y),i.className="selector",n.className="wall-filter",z.className="row",P.className="wall-cards",J.className="east-nav",R.className="tab-pane fade show active",R.id="contribute",g(R,"role","tabpanel"),g(R,"aria-labelledby","contribute-tab"),V.className="tab-pane fade",V.id="about",g(V,"role","tabpanel"),g(V,"aria-labelledby","about-tab"),X.className="tab-pane fade",X.id="support",g(X,"role","tabpanel"),g(X,"aria-labelledby","support-tab"),Q.className="tab-content",Q.id="pills-tabContent",I.className="container",G.className="wall-misc",a.id="wall"},m:function(t,e){y(a,t,e),N(n,a),N(i,n),N(s,i),N(c,s),N(u,c),N(f,s),N(d,f),N(h,s),N(k,h),N(S,s),N(q,S),N(A,s),N(O,A),N(j,s),N(E,j),N(L,s),N(M,L),N(T,s),N($,T),N(B,s),N(F,B),N(H,a),N(P,a),N(z,P);for(var r=0;r<tt.length;r+=1)tt[r].m(z,null);N(D,a),N(G,a),N(I,G),N(J,I),N(K,I),N(Q,I),N(R,Q),at._mount(R,null),N(U,Q),N(V,Q),nt._mount(V,null),N(W,Q),N(X,Q),rt._mount(X,null)},p:function(e,a){var n=a.filteredCards;if(e.filteredCards){for(var r=0;r<n.length;r+=1){var i=v({},a,{card:n[r],card_index:r});tt[r]?tt[r].p(e,i):(tt[r]=_(t,i),tt[r].c(),tt[r].m(z,null))}for(;r<tt.length;r+=1)tt[r].u(),tt[r].d();tt.length=n.length}},u:function(){w(a);for(var t=0;t<tt.length;t+=1)tt[t].u()},d:function(){C(s,"change",Y),x(tt),at.destroy(!1),nt.destroy(!1),rt.destroy(!1)}}}function _(t,e){var n,r=e.card,i=(e.card_index,new a.default({root:t.root,data:{card:r}}));return{c:function(){n=p("div"),i._fragment.c(),this.h()},h:function(){n.className="col-xl-4 col-lg-6 col-md-12 col-sm-12"},m:function(t,e){y(n,t,e),i._mount(n,null)},p:function(t,e){r=e.card,e.card_index;var a={};t.filteredCards&&(a.card=r),i._set(a)},u:function(){w(n)},d:function(){i.destroy(!1)}}}function h(t){k(this,t),this._state=v(this.store._init(["cards"]),f(),t.data),this.store._add(this,["cards"]),this._recompute({$cards:1,filter:1},this._state),this._handlers.destroy=[S],t.root||(this._oncreate=[],this._beforecreate=[],this._aftercreate=[]),this._fragment=d(this,this._state),t.target&&(this._fragment.c(),this._fragment.m(t.target,t.anchor||null),this._lock=!0,q(this._beforecreate),q(this._oncreate),q(this._aftercreate),this._lock=!1)}function v(t){for(var e,a,n=1,r=arguments.length;n<r;n++)for(e in a=arguments[n])t[e]=a[e];return t}function p(t){return document.createElement(t)}function b(t){return document.createTextNode(t)}function m(t,e,a){t.addEventListener(e,a,!1)}function g(t,e,a){t.setAttribute(e,a)}function y(t,e,a){e.insertBefore(t,a)}function N(t,e){e.appendChild(t)}function w(t){t.parentNode.removeChild(t)}function C(t,e,a){t.removeEventListener(e,a,!1)}function x(t){for(var e=0;e<t.length;e+=1)t[e]&&t[e].d()}function k(t,e){t._observers={pre:H(),post:H()},t._handlers=H(),t._bind=e._bind,t.options=e,t.root=e.root||t,t.store=t.root.store||e.store}function S(){this.store._remove(this)}function q(t){for(;t&&t.length;)t.shift()()}function A(t){this.destroy=P,this.fire("destroy"),this.set=this.get=P,!1!==t&&this._fragment.u(),this._fragment.d(),this._fragment=this._state=null}function O(t){return t?this._state[t]:this._state}function j(t,e){var a=t in this._handlers&&this._handlers[t].slice();if(a)for(var n=0;n<a.length;n+=1)a[n].call(this,e)}function E(t,e,a){var n=a&&a.defer?this._observers.post:this._observers.pre;return(n[t]||(n[t]=[])).push(e),a&&!1===a.init||(e.__calling=!0,e.call(this,this._state[t]),e.__calling=!1),{cancel:function(){var a=n[t].indexOf(e);~a&&n[t].splice(a,1)}}}function L(t,e){if("teardown"===t)return this.on("destroy",e);var a=this._handlers[t]||(this._handlers[t]=[]);return a.push(e),{cancel:function(){var t=a.indexOf(e);~t&&a.splice(t,1)}}}function M(t){this._set(v({},t)),this.root._lock||(this.root._lock=!0,q(this.root._beforecreate),q(this.root._oncreate),q(this.root._aftercreate),this.root._lock=!1)}function T(t){var e=this._state,a={},n=!1;for(var r in t)this._differs(t[r],e[r])&&(a[r]=n=!0);n&&(this._state=v({},e,t),this._recompute(a,this._state),this._bind&&this._bind(a,this._state),this._fragment&&(z(this,this._observers.pre,a,this._state,e),this._fragment.p(a,this._state),z(this,this._observers.post,a,this._state,e)))}function $(t,e){this._fragment.m(t,e)}function B(){this._fragment&&this._fragment.u()}function F(e,a){return e!=e?a==a:e!==a||e&&"object"===(void 0===e?"undefined":t(e))||"function"==typeof e}function H(){return Object.create(null)}function P(){}function z(t,e,a,n,r){for(var i in e)if(a[i]){var o=n[i],s=r[i],l=e[i];if(l)for(var c=0;c<l.length;c+=1){var u=l[c];u.__calling||(u.__calling=!0,u.call(t,o,s),u.__calling=!1)}}}v(h.prototype,{destroy:A,get:O,fire:j,observe:E,on:L,set:M,teardown:A,_set:T,_mount:$,_unmount:B,_differs:F}),h.prototype._recompute=function(t,e){(t.$cards||t.filter)&&this._differs(e.filteredCards,e.filteredCards=u(e.$cards,e.filter))&&(t.filteredCards=!0)},exports.default=h;
+},{"./Card.svelte":32,"./Contribute.svelte":34,"./Support.svelte":36,"./About.svelte":35}],10:[function(require,module,exports) {
+module.exports="4ad85fcd9c9e0f3204080d5aae13de06.png";
+},{}],11:[function(require,module,exports) {
+module.exports="4a0057b9aa02aa3c759ba5754e213c8d.jpg";
+},{}],13:[function(require,module,exports) {
+module.exports="077b0fdd4970e892246db21123ae3ef6.png";
+},{}],12:[function(require,module,exports) {
+module.exports="cdf63d0c6cf9ee440ad63554f402e4dc.png";
+},{}],14:[function(require,module,exports) {
+module.exports="fc5b3d317a6da5d2fd3f705da792e304.png";
+},{}],15:[function(require,module,exports) {
+module.exports="1d21066bbd07ce78e17547bb7f52e5a0.png";
+},{}],16:[function(require,module,exports) {
+module.exports="e552e812560812e424cbf571b6202b82.png";
+},{}],18:[function(require,module,exports) {
+module.exports="b06069e74aef13310a03f79b433d015e.jpg";
+},{}],17:[function(require,module,exports) {
+module.exports="eef1b090d732646b2965b554647fe16d.jpg";
+},{}],19:[function(require,module,exports) {
+module.exports="2f4288cbb40e79b03bfe026fd636582f.jpg";
+},{}],20:[function(require,module,exports) {
+module.exports="572c8c3738bc125668ecde0323b7d1cd.jpg";
+},{}],21:[function(require,module,exports) {
+module.exports="a9f657a584d80d9dd8662b53bc9f609d.png";
+},{}],22:[function(require,module,exports) {
+module.exports="a33d89f2a5940e380e45e5de3b39a63c.png";
+},{}],23:[function(require,module,exports) {
+module.exports="66e37bc43a549ece973eb1a2ae273694.jpg";
+},{}],24:[function(require,module,exports) {
+module.exports="8e8d9eed9ee2520d89012295bf092577.png";
+},{}],25:[function(require,module,exports) {
+module.exports="44fa25923f45216f542ee89ebb1e5ea7.png";
+},{}],26:[function(require,module,exports) {
+module.exports="1349bc3a910784f4d5b396410e8ca311.png";
+},{}],28:[function(require,module,exports) {
+module.exports="bba7555a09ca534c5a5422725560e955.jpg";
+},{}],27:[function(require,module,exports) {
+module.exports="96f1ac8ad356f2023b19a80681420ace.png";
+},{}],29:[function(require,module,exports) {
+module.exports="7b57fb916f7684f6f2a85626708470de.png";
+},{}],30:[function(require,module,exports) {
+module.exports="846b376f8dfac01a9421f3aa02aa4a8f.svg";
+},{}],31:[function(require,module,exports) {
+module.exports="e5ab34d5143051870b22087c62cfab94.svg";
+},{}],6:[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=[{id:"google-classroom",name:"Google Classroom",url:"https://classroom.google.com",tagline:"Classroom allows educators to manage classes, set assignments and interact with students all in one place.",description:"Classroom connects students and instructors using tools such as Google Docs, Calendar, Drive, and Forms to build a streamlined platform for learning. Teachers can create classes, distribute assignments, communicate and stay organized whilst students can see their assignments on their To-Do page, share resources and interact in the class stream.",colour:"#FDBC45",assets:{logo:{styles:"",image:require("./img/google_classroom.png")},cover:{styles:"background-size: cover !important;",image:require("./img/classroom_screenshot.jpg")}},activities:["connect","collaborate","curate"],uses:["Assessment","Discussion","Feedback"],submitter:{name:"A Rhead",title:"Learning Developer"},videoURL:"https://www.youtube.com/embed/UPgnim0Q5cs?rel=0&amp;showinfo=0?&cc_load_policy=1",framework:{enhance:"Set up a class and keep it organised by applying topics to posts. Structure materials by week, lesson, theme or assignment.",empower:"Add learning resources in Google formats, amending sharing settings for students to edit or add comments.",extend:"Use the Classroom stream to set up discussion topics or create quick polls that engage students inbetween lessons."}},{id:"gdocs",name:"Google Docs",url:"https://www.google.com/docs/about",tagline:"An online word processor that allows you to write, edit and collaborate wherever you are.",description:"As part of GSuite (formerly Google Apps for Education), Google Docs is a simple but powerful word processing application that helps to create documents from any device. As well as familiar features for formatting and styling, Docs makes it possible to work together in real-time. Sharing files is easy too, with adjustable security settings that make emailing multiple copies of the same document a thing of the past.",colour:"#5793EF",assets:{logo:{styles:"background-color: #343A40; background-size: 20px; background-position: 50% 50%;",image:require("./img/docs.png")},cover:{styles:"background-color: #fff;",image:require("./img/docs_background.png")}},activities:["create","collaborate"],uses:["Collaborative writing","Feedback","Workbooks"],submitter:{name:"M Davys",title:"Senior Teaching Fellow in Law"},videoURL:"https://www.youtube.com/embed/ns0U3zkHG7w?rel=0&amp;showinfo=0?&cc_load_policy=1",framework:{enhance:"Convert a resource that would normally be given to students in a paper or static electronic format into a Google Doc and share it for students to annotate and make their own.",empower:"Organise an 'edit-a-thon' where groups of students are encouraged to collaborate in real-time on a Google Doc to demonstrate their knowledge of a particular topic.",extend:"Ask students to share their Google Doc and provide feedback by annotating and inserting comments that prompt further discussion or revisions to their original work."}},{id:"medium",name:"Medium",url:"https://medium.com/",tagline:"A blogging platform that makes writing and sharing posts simple.",description:"Medium simplifies the writing process, producing great looking articles that can easily be shared online. Author your own or show your appreciation for other posts by using the applaud feature, or bookmark articles to read later. Medium allows anybody to read, write and respond to stories, with licensing options to allow others to remix and reuse your posts. Found an article that needs more discussion? Write a reponse underneath or highlight specific text and add comments inline.",colour:"#000",assets:{logo:{styles:"background-color: #000; background-size: 30px; background-position: 50% 60%; border: 3px solid #000;",image:require("./img/medium.png")},cover:{styles:"",image:require("./img/medium_logo.png")}},activities:["create","check"],uses:["Blogging","Discussion","Portfolios","Reflection"],submitter:{name:"T Maurice",title:"Digital Communications Officer"},videoURL:"https://www.youtube.com/embed/U0QOonYmDuU?rel=0&amp;showinfo=0?&cc_load_policy=1",framework:{enhance:"Start a module blog, writing occasional articles that summarise topics or respond to a related post from another source or Medium author.",empower:"Encourage students to respond to your article, asking questions inline or offering their own thoughts. Responses do not have to made public.",extend:"Set a reflective task that asks students to write their own posts, either keeping to the same topic or setting a new theme. Add their posts to a class publication."}},{id:"mentimeter",name:"Mentimeter",url:"https://www.mentimeter.com",tagline:"An online polling tool that encourages audience participation.",description:"Energise presentations and make lectures more interactive by building informal quizzes and polls that students can respond to using any Internet connected device. Live results are displayed on-screen or hidden until the presenter chooses to reveal them. Great for formative feedback and to monitor understanding.",colour:"#00a5bd",assets:{logo:{styles:"background-color: #00a5bd",image:require("./img/mentimeter.png")},cover:{styles:"",image:require("./img/mentimeter_logo.jpg")}},activities:["create","captivate","check"],uses:["Assessment","Polls","Quizzes","Feedback"],submitter:{name:"Dr C Little",title:"Learning Developer"},videoURL:"https://www.youtube.com/embed/f1s7UhcIA7I?rel=0&amp;showinfo=0?&cc_load_policy=1",framework:{enhance:"Plan an icebreaker activity that poses a quick quiz to help gauge students prior knowledge and to introduce topics that are due to be covered.",empower:"Integrate your slides with Mentimeter and use Reactions to help indicate when students are ready to move on or don't understand.",extend:"Create a short answer question that asks students to suggest revision topics. Turn the answers into a poll and use it to inform the next session's content."}},{id:"office-lens",name:"Office Lens",url:"https://twitter.com/msofficelens",tagline:"Capture images of whiteboards and documents to make them into readable learning resources.",description:"Simply put, Office Lens puts a scanner in your pocket. Frame your shot, shoot from any angle, and like magic, it digitises notes from whiteboards or physical documents for you to easily tidy up. Office Lens uses optical character recognition (OCR) to recognise and convert static images into searchable, editable documents such as Word or PDF. Never lose your sticky notes again!",colour:"#D83C18",assets:{logo:{styles:"background-size: 40px; background-position: 50% 20%;",image:require("./img/officelens.jpg")},cover:{styles:"background-size: cover !important;",image:require("./img/officelens_background.jpg")}},activities:["create","capture"],uses:["Notetaking","Annotating","Documenting"],submitter:{name:"K Beaumont",title:"Learning Developer"},videoURL:"https://www.youtube.com/embed/s5_giVIUWr4?rel=0&amp;showinfo=0?&cc_load_policy=1",framework:{enhance:"Share images of whiteboard notes with students after class, or encourage them to capture their own and use as a revision resource later on.",empower:"Encourage students to take pictures of their written notes and use them as a starting point for wider research and to add to more detailed, typed notes.",extend:"Using their captured notes, ask students to create quizzes and test each other's knowledge and understanding of topics covered in class."}},{id:"padlet",name:"Padlet",url:"https://www.padlet.com",tagline:"Padlet creates digital noticeboards that make sharing and collaborating easy.",description:"Padlet provides a flexible digital canvas to add a variety of content. Drag in a video, snap a picture, write text posts or upload documents! Make it even more individual by choosing custom wallpapers and themes. Padlets can be selectively shared and edited among multiple contributors, in real time.",colour:"#FF4A76",assets:{logo:{styles:"background-color: #202338;",image:require("./img/padlet.jpg")},cover:{styles:"",image:require("./img/padlet_logo.png")}},activities:["create","collaborate","curate"],uses:["Mindmapping","Presentation","Discussion"],submitter:{name:"T Hinchcliffe",title:"Head of Student Learning & Development"},videoURL:"https://www.youtube.com/embed/P9CJauuNw8c?rel=0&amp;showinfo=0?&cc_load_policy=1",framework:{enhance:"Ask students to engage with a lecture topic by adding their thoughts and questions onto a Padlet wall, forming the basis for a discussion.",empower:"Pre-load a collection of resources onto a Padlet wall and ask students to critique them, sharing their responses as comments underneath.",extend:"Collect feedback and take questions from students by maintaining a Padlet wall throughout the semester. Tidy and share it with future cohorts."}},{id:"pinterest",name:"Pinterest",url:"https://www.pinterest.com",tagline:"Upload, save and manage visual content (e.g. images and video) via collections known as pinboards.",description:"Pinterest is a visual bookmarking tool that helps to discover and save creative ideas. Known as pins, images and videos can be uploaded, saved and categorised into collections  (pinboards) that can be kept private or shared online. Pinboards can be split into sections and made collaborative, allowing others to add content. Pinterest's bookmarking tool can also be used to quickly save content from around the web.",colour:"#BD081C",assets:{logo:{styles:"background-color: #BD081C; background-size: 30px; background-position: 50% 25%;",image:require("./img/pinterest.png")},cover:{styles:"background-size: cover !important;",image:require("./img/pinterest_logo.jpg")}},activities:["create","collaborate","curate"],uses:["Bookmarking","Collating","Researching"],submitter:{name:"Dr E de Quincey",title:"Senior Lecturer in Computing"},videoURL:"https://www.youtube.com/embed/qFdxRkYldT8?rel=0&amp;showinfo=0?&cc_load_policy=1",framework:{enhance:"Use a pinboard to curate and share visual content with students. Collate media relevant to the module/topic being taught.",empower:"Open up the module pinboard for contributions and ask students to pin media they think relevant. Ask them to discuss/critique their choices.",extend:"In groups, encourage students to create a collaborative pinboard. Assess it as a virtual poster, requiring a narrative that explains their choices and how it relates to a particular topic."}},{id:"pocket",name:"Pocket",url:"https://www.getpocket.com",tagline:"A social bookmarking tool that saves web content across all of your devices.",description:"Pocket saves and organises interesting articles, videos and other web content so you can return to it later. Once saved, your personalised library is available on any device, from anywhere — even offline. Easily share items via email or post them to Twitter and Facebook for others to discover.",colour:"#EF3E56",assets:{logo:{styles:"background-size: 40px; background-position: 50% 30%; border: solid 3px #fff;",image:require("./img/pocket.png")},cover:{styles:"",image:require("./img/pocket_logo.png")}},activities:["curate"],uses:["Bookmarking","Collating","Researching"],submitter:{name:"Dr J Herbert",title:"Senior Lecturer in Politics"},videoURL:"https://biteable.com/watch/embed/video-placeholder-1602227/9a0c58e895f0d2a844c91c1718fb40c2786b47d1?autoplay=1",framework:{enhance:"Begin building a reading list by installing Pocket on your devices, including the add-on available for most modern web browsers (e.g. Google Chrome).",empower:"Start using tags and favourites to organise your list. Tags can be applied to multiple items to build categorised sub-lists.",extend:"Set up a module Twitter account or use a hashtag (e.g. #HUM1020) to start sharing articles with students from your Pocket reading list."}},{id:"snagit",name:"Snagit",url:"http://www.snagit.com",tagline:"Capture screenshots or record video to visually explain anything from your screen.",description:"Snagit is an all-in-one tool for image and video capture, making it easy to create a variety of learning resources. For example, produce step-by-step tutorials, short introductory videos or annotated images. Snagit's integrated editor simplifies the process by providing a range of professional markup options.",colour:"#2C72B7",assets:{logo:{styles:"background-color: #2C72B7; background-size: 32px; background-position: 50% 25%; border: solid 3px #fff;",image:require("./img/snagit.png")},cover:{styles:"background-size: cover !important;",image:require("./img/snagit_logo.jpg")}},activities:["create","capture"],uses:["Screencasting","Annotating","Editing"],submitter:{name:"E Tennant",title:"English Language Teaching Fellow"},videoURL:"https://biteable.com/watch/embed/video-placeholder-1602227/9a0c58e895f0d2a844c91c1718fb40c2786b47d1?autoplay=1",framework:{enhance:"Record and upload a short video that introduces yourself, the topics you'll be covering and what assessments to expect.",empower:"Narrate a recommended text (e.g. journal article), identifying key points to promote skills that students can apply in their wider reading.",extend:"Use Snagit to record audiovisual feedback based on anonymised draft submissions to help resolve common assignment issues."}},{id:"socrative",name:"Socrative",url:"https://www.socrative.com",tagline:"An app that encourages student participation via online polls and quizzes for real time feedback.",description:"Engage students with Socrative, an app that helps to build activities that provide feedback as learning happens. Launch pre-prepared quizzes or ask ad-hoc questions to quickly assess students and get immediate insights into their understanding at class, individual or question level. Automatically populated results visualise feedback, either to be shared with the class or collected anonymously by the tutor.",colour:"#8EB4CF",assets:{logo:{styles:"background-color: #ffffff; background-size: 30px; background-position: 50% 60%; border: 3px solid #ffffff;",image:require("./img/socrative.png")},cover:{styles:"",image:require("./img/socrative_logo.png")}},activities:["create","captivate","check"],uses:["Assessment","Polls","Quizzes","Feedback"],submitter:{name:"Dr R Leach",title:"Head of Social Science and Public Policy"},videoURL:"https://biteable.com/watch/embed/video-placeholder-1602227/9a0c58e895f0d2a844c91c1718fb40c2786b47d1?autoplay=1",framework:{enhance:"Design a short quiz using different question types to informally check students understanding during a lecture or seminar.",empower:"Ask students to work in pairs or small groups to discuss their answers before responding to the quiz. Use a 'Space Race' for added competitiveness.",extend:"Download the feedback report and identify areas which students had difficulty with. Use it to recap topics or inform future learning design."}},{id:"zeetings",name:"Zeetings",url:"https://www.zeetings.com",tagline:"Seamlessly combine slides, multimedia content, polls and Q&A features into interactive presentations.",description:"Zeetings helps to create engaging presentations by combining a range of interactive tools into one, easy to use interface. Add slides from existing decks or use Zeetings to create brand new presentations, quickly embedding videos, polls and surveys. A Zeeting is interactive, so share with students to receive real time feedback and live questions. With in-built analytics, responses are saved automatically to repurpose outside of the classroom.",colour:"#2BA0D8",assets:{logo:{styles:"background-color: #343A40; background-size: 20px; background-position: 50% 50%;",image:require("./img/zeetings.svg")},cover:{styles:"",image:require("./img/zeetings_logo.svg")}},activities:["create","captivate","check"],uses:["Presentations","Notetaking","Polls","Feedback"],submitter:{name:"J Preston",title:"Lecturer in Midwifery"},videoURL:"https://www.youtube.com/embed/qKsY_083uOk?rel=0&amp;showinfo=0?&cc_load_policy=1",framework:{enhance:"Import a previous presentation (e.g. PowerPoint) and add some interactive slides that will test student's comprehension during the session.",empower:"Introduce students to the Q&A feature. Ask them to pose questions anonymously and vote on which ones they think should be answered first.",extend:"Use Zeetings' analytics feature to export all questions and allocate a revision task that requires students to present back an answer at the next class."}}];
+},{"./img/google_classroom.png":10,"./img/classroom_screenshot.jpg":11,"./img/docs.png":13,"./img/docs_background.png":12,"./img/medium.png":14,"./img/medium_logo.png":15,"./img/mentimeter.png":16,"./img/mentimeter_logo.jpg":18,"./img/officelens.jpg":17,"./img/officelens_background.jpg":19,"./img/padlet.jpg":20,"./img/padlet_logo.png":21,"./img/pinterest.png":22,"./img/pinterest_logo.jpg":23,"./img/pocket.png":24,"./img/pocket_logo.png":25,"./img/snagit.png":26,"./img/snagit_logo.jpg":28,"./img/socrative.png":27,"./img/socrative_logo.png":29,"./img/zeetings.svg":30,"./img/zeetings_logo.svg":31}],2:[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0});var e=function(){function e(e,t){for(var r=0;r<t.length;r++){var n=t[r];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}return function(t,r,n){return r&&e(t.prototype,r),n&&e(t,n),t}}(),t=require("svelte/store"),r=require("./templates/App.svelte"),n=u(r),a=require("./templates/Wall.svelte"),o=u(a),i=require("./cards.js"),c=u(i);function u(e){return e&&e.__esModule?e:{default:e}}function s(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function l(e,t){if(!e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return!t||"object"!=typeof t&&"function"!=typeof t?e:t}function f(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function, not "+typeof t);e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,enumerable:!1,writable:!0,configurable:!0}}),t&&(Object.setPrototypeOf?Object.setPrototypeOf(e,t):e.__proto__=t)}var d=function(r){function n(){return s(this,n),l(this,(n.__proto__||Object.getPrototypeOf(n)).apply(this,arguments))}return f(n,t.Store),e(n,[{key:"loadRandomCard",value:function(e){var t=this.get("cards").filter(function(t){return t.id!==e}),r=t[Math.floor(Math.random()*t.length)];this.set({currentCard:r.id})}}]),n}();c.default.forEach(function(e){var t=[{name:"create",description:"Build learning resources, from scratch or remixing existing materials."},{name:"captivate",description:"Create interactive learning opportunities that engage."},{name:"check",description:"Gauge students understanding."},{name:"connect",description:"Communicate with students in real time or asynchronously."},{name:"collaborate",description:"A platform for collaboration and co-creation."},{name:"curate",description:"Collect, organise and share content."},{name:"capture",description:"Record an event, activity or artefact."}];e.activities=e.activities.map(function(e){return t.find(function(t){return t.name===e})})});var p=new d({cards:c.default,currentCard:c.default[Math.floor(Math.random()*c.default.length)].id,currentPage:"card"}),m=new URL(document.location).searchParams,h=m.get("display"),y="wall"===h?o.default:n.default,b=new y({target:document.body,store:p});exports.default=b;
+},{"svelte/store":37,"./templates/App.svelte":7,"./templates/Wall.svelte":8,"./cards.js":6}]},{},[2])
