--- 58c09e6be93a8ffa10bb5161944f0090.css
@@ -1,257 +1,0 @@
-body {
-	margin: 0;
-	font-size: 14px;
-	background-color: #ffffff;
-}
-
-/* NAVIGATION */
-
-.card-header {
-    background-color: #fff;
-}
-
-.card-header a {
-    color: black;
-    font-size: 1.1em;
-}
-
-.nav-item {
-    /* width: 60px; */
-    text-align: center;
-}
-
-iframe {
-    margin-top: 5px;
-}
-
-/* Text Styles */
-
-h4 {
-    margin-left: 25px;
-    font-family: 'Catamaran', sans-serif;
-    font-weight: 700;
-    margin-bottom: 20px;
-}
-
-p {
-    font-family: 'Catamaran', sans-serif;
-    font-weight: 500;
-}
-
-a {
-    font-family: 'Catamaran', sans-serif;
-    font-weight: 700;
-}
-
-select {
-    -webkit-appearance: none;
-    -moz-appearance: none;
-    appearance: none;
-    background: url(336679ccf73455e201ce2c0e1ef3815b.png) right 10px center / 30px no-repeat #fff;
-    padding-right: 20px;
-}
-
-/* MAIN BODY */
-
-body {
-	font-size: 14px;
-	font-family: "Helvetica Nueue", Arial, Verdana, sans-serif;
-	background-color: #ffffff;
-}
-
-iframe {
-    width: 100%;
-}
-
-/* CATALOGUE TAB */
-
-.bs-callout {
-    margin: 15px 0px 10px 0px;
-    padding: 20px 30px 5px 20px;
-    border-left: 5px solid #eee;
-}   
-
-.bs-callout-catalogue {
-    border-color: #fccc30;
-    background-color: #fafafa;
-}
-
-.bs-callout .icon {
-    float: left;
-    font-size: 1.5em;
-    margin-top: -4px;
-}
-
-.bs-callout h4 {
-    margin-left: 35px;
-}
-
-.bs-callout p {
-    margin-left: 35px;
-}
-
-.dropdown-toggle {
-    background-color: #343a40;
-    color: #fff;
-    margin-top: 10px;
-    margin-bottom: 10px;
-    font-family: 'Catamaran', sans-serif;
-    font-weight: 500;
-}
-
-.dropdown-toggle:hover {
-    background-color: #343a40;
-    color: #fff;
-}
-
-.dropdown-toggle:focus {
-    background-color: #343a40;
-    color: #fff;
-}
-
-.cataloguepane {
-    height: 500px;
-    overflow-y: scroll;
-    margin-bottom: 15px;
-}
-
-.applist {
-    margin-top: 15px;
-}
-
-.submitter {
-    margin-top: -10px;
-}
-
-.appinfo {
-    margin-top: -10px;
-}
-
-/* Pills */
-
-.create {
-    background-color: #a168c8;
-}
-
-.connect {
-    background-color: #4db6ac;
-}
-
-.collaborate {
-    background-color: #f06292;
-}
-
-.curate {
-    background-color: #fccc30;
-}
-
-.capture {
-    background-color: #4fc3f7;
-}
-
-.captivate {
-    background-color: #4b768d;
-}
-
-.check {
-    background-color: #dadfe1;
-}
-
-
-/* CONTRIBUTE */
-
-
-/* Main Body */
-
-.bs-callout-contribute {
-    background-color: #fafafa;
-    border-color: #a168c8;
-}
-
-.bs-thanks {
-    display: none;
-    background-color: #fafafa;
-    border-color: #4db6ac;
-}
-
-.btn-block:hover {
-    cursor: pointer;
-}
-
-form input:focus,
-form textarea:focus {
-    border-color: #a168c8 !important;
-    border-width: 1px;
-}
-
-/* Text */
-
-.instruction {
-    color: #90A4AD;
-}
-
-/* Form */
-
-form {
-    margin-top: 20px;
-}
-
-.form-group label {
-    font-family: 'Catamaran', sans-serif;
-    font-weight: 500;
-}
-
-.submitbutton {
-    margin-bottom: 15px;
-    background-color: #343a40;
-    color: #fff;
-}
-
-.submitbutton:hover {
-    opacity: 0.9;
-}
-
-/* Honeypot */
-
-#honeypot {
-  display: none; 
-}
-
-
-/* ABOUT */
-
-
-.bs-callout-about {
-    background-color: #fafafa;
-    border-color: #4dd0e1;
-    margin-bottom: 15px;
-}
-
-.bs-callout a {
-    color: #4dd0e1;
-}
-
-.creativecommons {
-    margin-bottom: 20px;
-    margin-left: 35px;
-}
-
-/* Clearfix */
-
-.clearfix:before,
-.clearfix:after {
-    content: " "; /* 1 */
-    display: table; /* 2 */
-}
-
-.clearfix:after {
-    clear: both;
-}
-
-/**
- * For IE 6/7 only
- * Include this rule to trigger hasLayout and contain floats.
- */
-.clearfix {
-    *zoom: 1;
-}
