--- 5eb6f38e3c9de02050f5d507b2cf5659.css
@@ -1,1 +1,0 @@
-#wall .wall-filter{background-color:#343a40!important;padding:1.2em;height:100px}#wall .wall-cards{padding:0 1em}#wall .wall-cards,#wall .wall-misc{height:calc(100vh - 100px);overflow-y:scroll;overflow-x:hidden}#wall .wall-misc{background-color:#343a40!important;width:360px;float:right}#wall .nav-item{margin:5px;border:1px solid #fff;border-radius:5px}#wall .nav-item a{color:#fff}#wall .nav-item:active,#wall .nav-item:hover{background:#000}#wall .nav-item.visited{background-color:transparent}#wall .contacts-nav .nav-pills .nav-link.active,#wall .east-nav .nav-item:hover,#wall .east-nav .nav-pills .nav-link.active{background-color:#000}#wall .details{border-bottom-left-radius:10px;border-bottom-right-radius:10px}#wall .nextcard{display:none}#wall .alert{font-family:Catamaran,sans-serif;font-size:1.15em}#wall .alert i{font-size:1em}#wall .alert button{color:#fff;font-family:system-ui,sans-serif;line-height:.8em}#wall .dropdown-toggle{margin-left:auto;margin-right:auto}#wall .bs-callout-contribute{background-color:#343a40;border-color:#ba68c8;color:#fff}#wall .bs-thanks{display:none;background-color:#343a40;border-color:#4db6ac;color:#fff}#wall .btn-block{font-family:Catamaran,sans-serif;font-weight:700}#wall form input:focus,#wall form textarea:focus{border-color:#ba68c8!important}#wall .form-group label{color:#fff}#wall .submitbutton{background-color:#ba68c8}#wall .bs-callout-support{background-color:#343a40;border-color:#fccc30;color:#fff}#wall .bs-callout-support .nav-item .nav-link{color:#fff}#wall .list-group-item{background-color:transparent;border:1px solid #fff;color:#fff!important}#wall .list-group-item:hover{background-color:#000}#wall .list-group-item i{color:#fccc30}#wall .bs-callout-about{background-color:#343a40;border-color:#4dd0e1;color:#fff}
