+++ iframe_wall.html
@@ -3,10 +3,6 @@
     <head>
         <meta name="viewport" content="width=device-width, initial-scale=1.0">   
         <style>
-
-            body {
-                margin: 0;
-            }
         
             @media only screen and (max-width: 200px) {
                 iframe {
@@ -17,12 +13,10 @@
             #topIframe {
                 margin: auto;
                 display: block;
-                height: 100vh;
-                width: 100vw;
             }
         </style>
     </head>
     <body>
-        <iframe id="topIframe" src="https://humsstel.github.io/tel_cards/?display=wall"  frameborder="0"></iframe>
+        <iframe id="topIframe" src="https://danieladamharding.github.io/tel_cards_test/?display=wall" width="100%" height="900px"  frameborder="0"></iframe>
     </body>
 </html>
